% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{bcbio}
\alias{bcbio<-}
\alias{metrics}
\alias{plotGene}
\alias{plotQC}
\title{S4 Generics}
\usage{
bcbio(object, ...)

bcbio(object, ...) <- value

metrics(object, ...)

plotGene(object, ...)

plotQC(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
Varies, depending upon the method.
}
\description{
S4 Generics
}
\keyword{internal}
