% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Assert Checks}
\arguments{
\item{severity}{How severe should the consequences of the assertion be?
Either \code{stop}, \code{warning}, or \code{message}.}
}
\description{
Assert Checks
}
\keyword{internal}
