% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_formal_interesting_groups}
\alias{assert_formal_interesting_groups}
\title{Interesting Groups Formal Assert Check}
\usage{
assert_formal_interesting_groups(x, interestingGroups, severity = "stop")
}
\arguments{
\item{x}{Object supporting \code{\link[=colnames]{colnames()}}, typically a \link{data.frame}.}

\item{interestingGroups}{Interesting groups character vector.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\description{
Prevent unwanted downstream behavior when a missing interesting group
is requested by the user.
}
\examples{
demultiplexed <- file.path(
    "http://bcbiobase.seq.cloud",
    "sample_metadata",
    "demultiplexed.xlsx")
meta <- readSampleMetadataFile(demultiplexed)
assert_formal_interesting_groups(meta, "genotype")
}
\keyword{internal}
