% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertFormalAnnotationCol}
\alias{assertFormalAnnotationCol}
\title{Assert Formal Annotation Column}
\usage{
assertFormalAnnotationCol(x, sampleData,
  severity = getOption("assertive.severity", "stop"))
}
\arguments{
\item{sampleData}{Sample data.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Silent on success, stop on failure.
}
\description{
Assert Formal Annotation Column
}
\examples{
x <- data.frame(
    "sample_1" = c(1L, 2L),
    "sample_2" = c(3L, 4L),
    row.names = c("gene_1", "gene_2"),
    stringsAsFactors = FALSE
)
sampleData <- data.frame(
    "genotype" = c("wt", "ko"),
    row.names = c("sample_1", "sample_2"),
    stringsAsFactors = TRUE
)
assertFormalAnnotationCol(x, sampleData)
}
\seealso{
Other Assert Check Functions: \code{\link{assertFormalInterestingGroups}}
}
\author{
Michael Steinbaugh
}
