% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bcbioBase-package}
\alias{bcbioBase}
\alias{bcbioBase-package}
\title{bcbioBase}
\description{
Base functions for bcbio R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/bcbiobase/}
  \item \url{https://github.com/hbc/bcbioBase/}
  \item Report bugs at \url{https://github.com/hbc/bcbioBase/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Authors:
\itemize{
  \item Rory Kirchner \email{roryk@alum.mit.edu} (\href{https://orcid.org/0000-0003-4814-5885}{ORCID})
  \item Lorena Pantano \email{lorena.pantano@gmail.com} (\href{https://orcid.org/0000-0002-3859-3249}{ORCID})
}

Other contributors:
\itemize{
  \item Harvard Chan Bioinformatics Core \email{bioinformatics@hsph.harvard.edu} [copyright holder, funder]
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
