% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{bcbioBaseCacheURL}
\alias{bcbioBaseCacheURL}
\title{Cache URL}
\format{An object of class \code{character} of length 1.}
\usage{
bcbioBaseCacheURL
}
\description{
Cache URL
}
\examples{
bcbioBaseCacheURL
}
\keyword{datasets}
