% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{bcbioBaseTestsUrl}
\alias{bcbioBaseTestsUrl}
\title{bcbioBase test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcbioBaseTestsUrl
}
\description{
bcbioBase test data URL
}
\examples{
bcbioBaseTestsUrl
}
\keyword{datasets}
