% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-convertGenesToSymbols.R
\docType{methods}
\name{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertGenesToSymbols,SummarizedExperiment-method}
\title{Convert Genes to Symbols}
\usage{
\S4method{convertGenesToSymbols}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{SummarizedExperiment}, with unique gene symbols as the rownames.
}
\description{
Convert Genes to Symbols
}
\examples{
# SummarizedExperiment ====
x <- convertGenesToSymbols(rse_bcb)
show(x)
head(rownames(x))
}
\seealso{
Other Data Functions: \code{\link{flatFiles}},
  \code{\link{gene2symbol}}, \code{\link{sampleDirs}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
