% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToDropbox.R
\name{copyToDropbox}
\alias{copyToDropbox}
\title{Copy Files to Dropbox}
\usage{
copyToDropbox(files, dir)
}
\arguments{
\item{files}{Local file paths.}

\item{dir}{Relative path of remote Dropbox directory.}
}
\value{
Invisibly return \link{list} of rdrop2 output.
}
\description{
Copy Files to Dropbox
}
\examples{
\dontrun{
files <- c("raw_counts.csv.gz", "tpm.csv.gz")
copyToDropbox(
    files,
    dir = file.path("researcher", "project", "results")
)
}
}
\author{
Michael Steinbaugh, Victor Barerra, John Hutchinson
}
