% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{sampleMetadata}
\alias{sampleMetadata<-}
\title{Deprecated Functions}
\usage{
sampleMetadata(object, ...)

sampleMetadata(object, ...) <- value
}
\value{
No value.
}
\description{
Deprecated Functions
}
\keyword{internal}
