% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-flatFiles.R
\docType{methods}
\name{flatFiles}
\alias{flatFiles}
\alias{flatFiles,SummarizedExperiment-method}
\title{Get Flat Files from S4 Object}
\usage{
\S4method{flatFiles}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{list}.
}
\description{
Prepare an unstructured list of information saved in the S4 object,
for improved archival data storage.
}
\examples{
\dontrun{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))
flatFiles(bcb) \%>\% names()
}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
