% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-general.R
\name{general}
\alias{general}
\title{General Arguments}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
General Arguments
}
\keyword{internal}
