% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{geom_label_bcbio}
\alias{geom_label_bcbio}
\title{Repulsive Textual Annotations}
\usage{
geom_label_bcbio(data = NULL, mapping = NULL, color = NULL, size = 4L)
}
\value{
\code{ggproto}.
}
\description{
Modified bcbio version of \code{\link[ggrepel:geom_label_repel]{ggrepel::geom_label_repel()}}. If advanced
customization of the text labels is required, simply use the ggrepel version
instead. This is a convenience function.
}
\examples{
x <- geom_label_bcbio()
class(x)
}
\seealso{
\code{\link[ggrepel:geom_label_repel]{ggrepel::geom_label_repel()}}.
}
\author{
Michael Steinbaugh
}
