% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromCommands.R
\name{getBarcodeCutoffFromCommands}
\alias{getBarcodeCutoffFromCommands}
\title{Get Cellular Barcode Cutoff from Commands Log}
\usage{
getBarcodeCutoffFromCommands(log)
}
\arguments{
\item{log}{\code{character}. Commands log.}
}
\description{
Get Cellular Barcode Cutoff from Commands Log
}
\author{
Michael Steinbaugh, Rory Kirchner
}
