% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromCommandsLog.R
\name{getBarcodeCutoffFromCommandsLog}
\alias{getBarcodeCutoffFromCommandsLog}
\title{Get Cellular Barcode Cutoff from Commands Log}
\usage{
getBarcodeCutoffFromCommandsLog(log)
}
\arguments{
\item{log}{\code{character}. Commands log.}
}
\description{
Get Cellular Barcode Cutoff from Commands Log
}
\author{
Michael Steinbaugh, Rory Kirchner
}
