% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGtfFileFromYaml.R
\name{getGtfFileFromYaml}
\alias{getGtfFileFromYaml}
\title{Get GTF file path from YAML}
\usage{
getGtfFileFromYaml(yaml)
}
\arguments{
\item{yaml}{\code{list}.
YAML metadata.}
}
\value{
\code{character(1)} or \code{NULL}.
File path if the file exists. \code{NULL} if the file does not exist.
}
\description{
Get GTF file path from YAML
}
\note{
Updated 2022-03-07.
}
\examples{
file <- file.path(bcbioBaseTestsUrl, "summary.yaml")
yaml <- import(file)
x <- getGtfFileFromYaml(yaml)
print(x)
}
\author{
Michael Steinbaugh
}
