% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{getLevelFromCommands}
\alias{getLevelFromCommands}
\title{Get Gene or Transcript Level from Commands Log}
\usage{
getLevelFromCommands(log)
}
\arguments{
\item{log}{\code{character}. Commands log.}
}
\value{
\code{string}.
}
\description{
Get Gene or Transcript Level from Commands Log
}
\author{
Michael Steinbaugh, Rory Kirchner
}
