% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricsFromYaml.R
\name{getMetricsFromYaml}
\alias{getMetricsFromYaml}
\title{Get quality control metrics from YAML}
\usage{
getMetricsFromYaml(yaml)
}
\arguments{
\item{yaml}{\code{list}.
YAML metadata.}
}
\value{
\code{DataFrame}.
}
\description{
Parse the summary YAML and return quality control metrics.
}
\note{
Metrics are only generated for a standard RNA-seq run with aligned
counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't
output the same metrics into the YAML.

Updated 2022-03-07.
}
\examples{
file <- file.path(bcbioBaseTestsUrl, "summary.yaml")
yaml <- import(file)
x <- getMetricsFromYaml(yaml)
summary(x)
colnames(x)
}
\author{
Michael Steinbaugh
}
