% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromCommands.R
\name{getUMITypeFromCommands}
\alias{getUMITypeFromCommands}
\title{Get UMI Type from Commands Log}
\usage{
getUMITypeFromCommands(log)
}
\arguments{
\item{log}{\code{character}. Commands log.}
}
\description{
Get UMI Type from Commands Log
}
\author{
Rory Kirchner, Michael Steinbaugh
}
