% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromCommandsLog.R
\name{getUMITypeFromCommandsLog}
\alias{getUMITypeFromCommandsLog}
\title{Get UMI Type from Commands Log}
\usage{
getUMITypeFromCommandsLog(log)
}
\arguments{
\item{log}{\code{character}. Commands log.}
}
\value{
\code{string}.
}
\description{
Get UMI Type from Commands Log
}
\seealso{
Other Data Functions: \code{\link{getBarcodeCutoffFromCommandsLog}},
  \code{\link{getLevelFromCommandsLog}},
  \code{\link{projectDir}}, \code{\link{runDate}},
  \code{\link{sampleDirs}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Data Functions}
