% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{ggplot2}
\alias{ggplot2}
\alias{bcbio_geom_abline}
\alias{bcbio_geom_label}
\alias{bcbio_geom_label_average}
\alias{bcbio_geom_label_repel}
\title{ggplot2 Functions}
\usage{
bcbio_geom_abline(xintercept = NULL, yintercept = NULL)

bcbio_geom_label(data = NULL, mapping = NULL, ...)

bcbio_geom_label_average(data, col, fun = c("median", "median"),
  digits = 0L, ...)

bcbio_geom_label_repel(data = NULL, mapping = NULL, color = NULL,
  size = 4L, ...)
}
\arguments{
\item{xintercept, yintercept}{\code{scalar numeric}. Specify either x- or y-axis
cutoff, but not both.}

\item{data}{\code{data.frame}.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{col}{\code{string}. Column.}

\item{fun}{\code{string}. Function name to use for average calculation. Currently
supports "\code{mean}" or "\code{median}".}

\item{digits}{\code{scalar integer}. Number of significant digits to use.
Defaults to rounded.}

\item{color}{\code{string}. Text color (e.g. \code{"orange"}).}

\item{size}{\code{scalar integer}. Font size.}
}
\value{
\code{ggproto}.
}
\description{
Convenience functions with modified defaults for
\href{http://ggplot2.org}{ggplot2}.
}
\section{bcbio_geom_abline}{


Horizontal or vertical cutoff line.
}

\section{bcbio_geom_label}{


Modified version of \code{\link[ggplot2:geom_label]{ggplot2::geom_label()}}.
}

\section{bcbio_geom_label_average}{


Add average labels to a plot. For example, \code{col} can be \code{nGene}. Note that
this function requires \code{sampleName} to be defined as a column. Median or mean
values are always calculated per sample (\code{sampleName}).
}

\section{bcbio_geom_label_repel}{


Repulsive textual annotations. Modified bcbio version of
\code{\link[ggrepel:geom_label_repel]{ggrepel::geom_label_repel()}}. If advanced customization of the text labels
is required, simply use the ggrepel version instead.
}

\examples{
# x-axis line
geom <- bcbio_geom_abline(xintercept = 1L)
geom

# y-axis line
geom <- bcbio_geom_abline(yintercept = 1L)
geom
geom <- bcbio_geom_label()
geom
\dontrun{
library(bcbioRNASeq)
data <- metrics(bcbioRNASeq::bcb_small)
geom <- bcbio_geom_label_average(
    data = data,
    col = "exonicRate",
    fun = "median"
)
geom
}
geom <- bcbio_geom_label_repel()
geom
}
\seealso{
\itemize{
\item \code{\link[ggplot2:geom_label]{ggplot2::geom_label()}}.
\item \code{\link[ggrepel:geom_label_repel]{ggrepel::geom_label_repel()}}.
}
}
\author{
Michael Steinbaugh
}
