% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDataVersions.R
\name{importDataVersions}
\alias{importDataVersions}
\title{Import data versions}
\usage{
importDataVersions(file)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{DataFrame}.
}
\description{
Import data versions
}
\note{
The \code{data_versions.csv} file is only generated for special genomes
containing additional information (e.g. the built-in \code{"hg38"} build).

Updated 2020-01-17.
}
\examples{
file <- file.path(bcbioBaseTestsURL, "data-versions.csv")
x <- importDataVersions(file)
print(x)
}
\author{
Michael Steinbaugh
}
