% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-interestingGroups.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups,SummarizedExperiment-method}
\alias{interestingGroups<-,SummarizedExperiment,character-method}
\title{Interesting Groups}
\usage{
\S4method{interestingGroups}{SummarizedExperiment}(object)

\S4method{interestingGroups}{SummarizedExperiment,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
Character vector.
}
\description{
Interesting Groups
}
\examples{
\dontrun{
load(system.file("extdata/se.rda", package = "bcbioBase"))

# SummarizedExperiment
interestingGroups(se)

# Assignment support
interestingGroups(se) <- "sampleID"
interestingGroups(se)
}
}
