% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{labelSep}
\alias{labelSep}
\title{Plot Label Separator}
\format{An object of class \code{character} of length 1.}
\usage{
labelSep
}
\description{
Plot Label Separator
}
\examples{
labelSep
}
\keyword{internal}
