% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{metadataBlacklist}
\alias{metadataBlacklist}
\title{Metadata Blacklist}
\format{An object of class \code{character} of length 9.}
\usage{
metadataBlacklist
}
\description{
Metadata Blacklist
}
\examples{
metadataBlacklist
}
\keyword{internal}
