% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimalSampleData.R
\name{minimalSampleData}
\alias{minimalSampleData}
\title{Minimal Sample Data}
\usage{
minimalSampleData(sample)
}
\arguments{
\item{sample}{Sample name (e.g. "description" in bcbio YAML).}
}
\value{
\code{data.frame}.
}
\description{
Minimal Sample Data
}
\examples{
minimalSampleData(c("sample 1", "sample 2"))
}
\seealso{
Other Data Functions: \code{\link{flatFiles}},
  \code{\link{sampleData}}, \code{\link{sampleDirs}},
  \code{\link{sampleNames}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
