% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotHeatmap-methods.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,SummarizedExperiment-method}
\title{Plot Heatmap}
\usage{
plotHeatmap(object, ...)

\S4method{plotHeatmap}{SummarizedExperiment}(object, interestingGroups,
  scale = c("row", "column", "none"), clusterRows = TRUE,
  clusterCols = TRUE, showRownames = FALSE, showColnames = TRUE,
  treeheightRow = 0L, treeheightCol = 50L, color = viridis,
  legendColor = NULL, borderColor = NULL, title = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments should be formatted in camel case, not snake case.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{scale}{\code{character} indicating whether the values should be centered and
scaled in either the row or column direction ("\code{row}", "\code{column}"), or
remain unscaled ("\code{none}").}

\item{clusterRows, clusterCols}{\code{logical} determining if rows or columns
should be arranged with hierarchical clustering. Alternatively, can define
an \code{hclust} object.}

\item{showRownames, showColnames}{Show row or column names.}

\item{treeheightRow, treeheightCol}{Size of the row and column dendrograms.
Use "\code{0}" to disable.}

\item{color}{Hexadecimal color function to use for plot. We recommend any of
these from the viridis package:
\itemize{
\item \code{\link[viridis:viridis]{viridis::viridis()}} (\emph{default}).
\item \code{\link[viridis:inferno]{viridis::inferno()}}.
\item \code{\link[viridis:magma]{viridis::magma()}}.
\item \code{\link[viridis:plasma]{viridis::plasma()}}.
}}

\item{legendColor}{Hexadecimal color function to use for legend labels.}

\item{borderColor}{\emph{Optional.} Border color. Disabled by default for
improved aesthetics.}

\item{title}{\emph{Optional.} Plot title.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
Construct a simple heatmap. By default, row- and column-wise hierarchical
clustering is performed using the Ward method, but this behavior can be
overrided by setting \code{clusterRows} or \code{clusterCols} to \code{FALSE}.
}
\examples{
# SummarizedExperiment ====
plotHeatmap(rse_dds, interestingGroups = "condition")

# Disable column clustering
plotHeatmap(rse_dds, clusterCols = FALSE)
}
\seealso{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.

Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotQuantileHeatmap}}
}
