% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotQuantileHeatmap-methods.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,SummarizedExperiment-method}
\title{Plot Heatmap with Quantile Breaks}
\usage{
plotQuantileHeatmap(object, ...)

\S4method{plotQuantileHeatmap}{SummarizedExperiment}(object, interestingGroups,
  ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments should be formatted in camel case, not snake case.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{n}{The number of breaks to create.}

\item{legend}{Show the color legend.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
Plot Heatmap with Quantile Breaks
}
\examples{
# SummarizedExperiment ====
plotQuantileHeatmap(rse_dds)

# Disable column clustering
plotQuantileHeatmap(rse_dds, clusterCols = FALSE)
}
\seealso{
Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotHeatmap}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
