% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSampleMetadata.R
\name{prepareSampleMetadata}
\alias{prepareSampleMetadata}
\title{Prepare Sample Metadata}
\usage{
prepareSampleMetadata(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{data.frame}.
}
\description{
Prepare Sample Metadata
}
