% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSummarizedExperiment.R
\name{prepareSummarizedExperiment}
\alias{prepareSummarizedExperiment}
\title{Prepare Summarized Experiment}
\usage{
prepareSummarizedExperiment(assays, rowRanges = NULL, colData = NULL,
  metadata = NULL, isSpike = NULL)
}
\arguments{
\item{assays}{List containing RNA-seq count matrices with matching
dimensions. Counts can be passed in either dense (\code{matrix}) or sparse
(\code{dgCMatrix}, \code{dgTMatrix}) format.}

\item{rowRanges}{\code{GRanges} describing assay rows. Must contain genomic
ranges. Can be left \code{NULL} if the genome is poorly annotated and/or ranges
aren't available from AnnotationHub.}

\item{colData}{\code{DataFrame} \code{data.frame}, or \code{matrix} describing assay
columns.}

\item{metadata}{\emph{Optional}. Metadata \code{list}.}

\item{isSpike}{Character vector of spike-in sequence rownames.}

\item{dropRows}{Drop unannotated rows (e.g. deprecated genes/transcripts). Only
applies when \code{rowRanges} is defined. Note that spike-ins declared in the
\code{isSpike} argument will not be dropped.}
}
\value{
\code{RangedSummarizedExperiment}.
}
\description{
This is a utility wrapper for \code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment()}}
that provides automatic subsetting for row and column data.
}
\details{
This function also provides automatic metadata slotting of multiple useful
environment parameters:
\itemize{
\item \code{date}: Today's date, returned from \code{\link[=Sys.Date]{Sys.Date()}}.
\item \code{wd}: Working directory, returned from \code{\link[=getwd]{getwd()}}.
\item \code{utilsSessionInfo}: [utils::sessionInfo().
\item \code{devtoolsSessionInfo}: [sessioninfo::session_info()].
}

[utils::sessionInfo().
\itemize{
\item \code{devtoolsSessionInfo}: [sessioninfo::session_info()]: R:utils::sessionInfo().%0A-%20%60devtoolsSessionInfo%60:%20[sessioninfo::session_info()
}
}
\examples{
\dontrun{
# Hello
}
}
