% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTemplate.R
\name{prepareTemplate}
\alias{prepareTemplate}
\title{Prepare R Markdown Template File}
\usage{
prepareTemplate(file = c("_output.yaml", "_footer.Rmd", "_header.Rmd",
  "_setup.R", "bibliography.bib"), sourceDir = system.file("rmarkdown/shared",
  package = "bcbioBase"))
}
\arguments{
\item{file}{File name.}

\item{sourceDir}{Source directory.}
}
\value{
No value.
}
\description{
If the required template dependency files aren't present, download latest
versions from the package website. \emph{Existing files are never overwritten.}
}
\details{
By default, this function will create local copies of these files:
\itemize{
\item \code{_footer.Rmd}
\item \code{_header.Rmd}
\item \code{_output.yaml}
\item \code{_setup.R}
\item \code{bibliography.bib}
}
}
\examples{
defaultFiles <- c(
    "_footer.Rmd",
    "_header.Rmd",
    "_output.yaml",
    "_setup.R",
    "bibliography.bib"
)

# Copy all of the default shared template files
prepareTemplate()
file.exists(defaultFiles)
unlink(defaultFiles)

# Request individual files
prepareTemplate("bibliography.bib")
file.exists("bibliography.bib")
unlink("bibliography.bib")

# Load the shared files from bcbioSingleCell
\dontrun{
prepareTemplate(
    sourceDir = system.file(
        "rmarkdown/shared",
        package = "bcbioSingleCell"
    )
)
}
}
\author{
Michael Steinbaugh
}
