% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioBase-package.R
\docType{data}
\name{projectDirectoryPattern}
\alias{projectDirectoryPattern}
\title{Project Directory Grep Pattern}
\format{An object of class \code{character} of length 1.}
\usage{
projectDirectoryPattern
}
\description{
Project Directory Grep Pattern
}
\keyword{internal}
