% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{projectDirPattern}
\alias{projectDirPattern}
\title{Project directory grep pattern}
\format{
An object of class \code{character} of length 1.
}
\usage{
projectDirPattern
}
\description{
Project directory grep pattern
}
\examples{
projectDirPattern
}
\keyword{datasets}
