% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataVersions.R
\name{readDataVersions}
\alias{readDataVersions}
\title{Read Data Versions}
\usage{
readDataVersions(file)
}
\arguments{
\item{file}{Data versions CSV file.}
}
\value{
\link{data.frame}.
}
\description{
Read Data Versions
}
\examples{
url <- paste(
    "http://bcbiobase.seq.cloud",
    "bcbio",
    "data_versions.csv",
    sep = "/")
readDataVersions(url) \%>\% glimpse()
}
\seealso{
Other File Utilities: \code{\link{readLogFile}}
}
