% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{readLog}
\alias{readLog}
\title{Read Log File}
\usage{
readLog(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
\code{character}.
}
\description{
Read Log File
}
\examples{
readLog("http://bcbiobase.seq.cloud/bcbio-nextgen.log") \%>\%
    head()
}
\author{
Michael Steinbaugh
}
