% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProgramVersions.R
\name{readProgramVersions}
\alias{readProgramVersions}
\title{Read Program Versions}
\usage{
readProgramVersions(file)
}
\arguments{
\item{file}{Program versions TXT file.}
}
\value{
\link{data.frame}.
}
\description{
Read Program Versions
}
\examples{
url <- paste(
    "http://bcbiobase.seq.cloud",
    "bcbio",
    "programs.txt",
    sep = "/")
readProgramVersions(url)
}
\seealso{
Other bcbio Utilities: \code{\link{prepareSummarizedExperiment}}
}
\keyword{internal}
