% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSampleData.R
\name{readSampleData}
\alias{readSampleData}
\title{Read Sample Metadata}
\usage{
readSampleData(file, lanes = 1L)
}
\arguments{
\item{file}{\code{string}. File path. Supports CSV, TSV, and XLSX file formats.}

\item{lanes}{\code{scalar integer}. Number of lanes used to split the samples into
technical replicates (\code{_LXXX}) suffix.}
}
\value{
\code{data.frame}.
}
\description{
Read Sample Metadata
}
\examples{
# Demultiplexed
x <- readSampleData("http://bcbiobase.seq.cloud/demultiplexed.csv")
glimpse(x)

# Multiplexed (e.g. inDrop single-cell RNA-seq)
x <- readSampleData("http://bcbiobase.seq.cloud/multiplexed.csv")
glimpse(x)
}
\author{
Michael Steinbaugh
}
