% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTx2gene.R
\name{readTx2gene}
\alias{readTx2gene}
\title{Transcript to Gene Annotations}
\usage{
readTx2gene(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
\code{tx2gene}.
}
\description{
Generates a \code{tx2gene} object containing \code{transcriptID} and \code{geneID} columns.
Rownames are set to the transcript IDs.
}
\note{
Doesn't attempt to strip transcript versions.
}
\examples{
x <- readTx2gene("http://bcbiobase.seq.cloud/tx2gene.csv")
print(x)
}
\seealso{
Other Read Functions: \code{\link{readDataVersions}},
  \code{\link{readLog}}, \code{\link{readProgramVersions}},
  \code{\link{readSampleData}}, \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
\concept{Read Functions}
