% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readYAML.R
\name{readYAML}
\alias{readYAML}
\alias{readYAMLSampleData}
\alias{readYAMLSampleMetrics}
\title{Read YAML Sample Data and Metrics}
\usage{
readYAMLSampleData(file)

readYAMLSampleMetrics(file)
}
\arguments{
\item{file}{\code{project-summary.yaml} file generated by bcbio.}
}
\value{
\code{data.frame}.
}
\description{
Read YAML Sample Data and Metrics
}
\note{
Metrics are only generated for a standard RNA-seq run with aligned
counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't
output the same metrics into the YAML.
}
\examples{
file <- "http://bcbiobase.seq.cloud/project-summary.yaml"
readYAMLSampleData(file) \%>\% glimpse()
readYAMLSampleMetrics(file) \%>\% glimpse()
}
\seealso{
Other Read Functions: \code{\link{readDataVersions}},
  \code{\link{readLog}}, \code{\link{readProgramVersions}},
  \code{\link{readSampleData}}, \code{\link{readTx2gene}}
}
\author{
Michael Steinbaugh
}
