% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rse_bcb}
\alias{rse_bcb}
\title{\code{RangedSummarizedExperiment} Coerced from \code{bcbioRNASeq}}
\format{An object of class \code{RangedSummarizedExperiment} with 500 rows and 6 columns.}
\usage{
rse_bcb
}
\description{
\code{RangedSummarizedExperiment} Coerced from \code{bcbioRNASeq}
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
