% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rse_small}
\alias{rse_small}
\title{\code{RangedSummarizedExperiment} Example}
\format{An object of class \code{RangedSummarizedExperiment} with 1000 rows and 12 columns.}
\usage{
rse_small
}
\description{
\code{RangedSummarizedExperiment} Example
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
