% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{runDate}
\alias{runDate}
\title{Run Date}
\usage{
runDate(projectDir)
}
\arguments{
\item{projectDir}{\code{string}. Project directory path.}
}
\description{
Get the run date from the project directory.
}
\examples{
runDate("2018-01-01_illumina_rnaseq")
}
\author{
Michael Steinbaugh
}
