% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDirs.R
\name{sampleDirs}
\alias{sampleDirs}
\title{Sample Directories}
\usage{
sampleDirs(uploadDir)
}
\arguments{
\item{uploadDir}{File path to bcbio run upload directory.}
}
\value{
Named \code{character} vector containing sample directory paths.
}
\description{
Sample Directories
}
\note{
Function will stop if no sample directories match.
}
\examples{
uploadDir <- system.file("extdata/bcbio", package = "bcbioBase")
sampleDirs(uploadDir)
}
\seealso{
Other Data Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{counts}}, \code{\link{flatFiles}},
  \code{\link{gene2symbol}},
  \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
