% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleNames.R
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames}
\title{Sample Names}
\usage{
sampleNames(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character} vector of the human readable sample names.
}
\description{
Requires that \code{sampleName} column is defined in \code{\link[=sampleData]{sampleData()}}.
}
\examples{
# SummarizedExperiment ====
sampleNames(rse_bcb)
}
\seealso{
Other Metadata Functions: \code{\link{sampleData}}
}
\author{
Michael Steinbaugh
}
