% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleYAML.R
\docType{methods}
\name{sampleYAML}
\alias{sampleYAML}
\alias{sampleYAML}
\alias{sampleYAML,list,character-method}
\title{Sample YAML Metadata Utilities}
\usage{
sampleYAML(yaml, keys, ...)

\S4method{sampleYAML}{list,character}(yaml, keys)
}
\arguments{
\item{yaml}{Project summary YAML \code{list}.}

\item{keys}{Nested operator keys, supplied as a character vector.}

\item{...}{Additional arguments.}
}
\value{
\code{tibble}.
}
\description{
Sample YAML Metadata Utilities
}
\note{
Metrics are only generated for a standard RNA-seq run with aligned
counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't
output the same metrics into the YAML.
}
\examples{
yaml <- basejump::readYAML(
    "http://bcbiobase.seq.cloud/project-summary.yaml"
)
sampleYAML(yaml, "metadata") \%>\% glimpse()
}
\seealso{
Other YAML Functions: \code{\link{sampleYAMLMetadata}},
  \code{\link{sampleYAMLMetrics}}
}
\author{
Michael Steinbaugh
}
