% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleYAMLMetadata.R
\docType{methods}
\name{sampleYAMLMetadata}
\alias{sampleYAMLMetadata}
\alias{sampleYAMLMetadata}
\alias{sampleYAMLMetadata,list-method}
\title{Sample Metadata from YAML}
\usage{
sampleYAMLMetadata(yaml, ...)

\S4method{sampleYAMLMetadata}{list}(yaml)
}
\arguments{
\item{yaml}{Project summary YAML \code{list}.}

\item{...}{Additional arguments.}
}
\value{
\code{tibble}.
}
\description{
Sample Metadata from YAML
}
\examples{
yaml <- basejump::readYAML(
    "http://bcbiobase.seq.cloud/project-summary.yaml"
)
sampleYAMLMetadata(yaml) \%>\% glimpse()
}
\seealso{
Other YAML Functions: \code{\link{sampleYAMLMetrics}},
  \code{\link{sampleYAML}}
}
