% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleYAMLMetrics.R
\docType{methods}
\name{sampleYAMLMetrics}
\alias{sampleYAMLMetrics}
\alias{sampleYAMLMetrics}
\alias{sampleYAMLMetrics,list-method}
\title{Sample Metrics from YAML File}
\usage{
sampleYAMLMetrics(yaml, ...)

\S4method{sampleYAMLMetrics}{list}(yaml)
}
\arguments{
\item{yaml}{Project summary YAML \code{list}.}

\item{...}{Additional arguments.}
}
\value{
\code{tibble}.
}
\description{
Sample Metrics from YAML File
}
\note{
bcbio doesn't calculate sample metrics when run in fast mode.
}
\examples{
yaml <- basejump::readYAML(
    "http://bcbiobase.seq.cloud/project-summary.yaml"
)
sampleYAMLMetrics(yaml) \%>\% glimpse()
}
\seealso{
Other YAML Functions: \code{\link{sampleYAMLMetadata}},
  \code{\link{sampleYAML}}
}
