% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-selectSamples.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,SummarizedExperiment-method}
\title{Select Samples}
\usage{
\S4method{selectSamples}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{SummarizedExperiment}.
}
\description{
Select Samples
}
\examples{
# SummarizedExperiment ====
x <- selectSamples(rse_bcb, day = 7L)
show(x)
colnames(x)
}
\seealso{
Other Data Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{flatFiles}}, \code{\link{gene2symbol}},
  \code{\link{sampleDirs}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
