% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-uniteInterestingGroups.R
\docType{methods}
\name{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\alias{uniteInterestingGroups}
\alias{uniteInterestingGroups,DataFrame-method}
\alias{uniteInterestingGroups,data.frame-method}
\alias{uniteInterestingGroups,tbl_df-method}
\title{Unite Interesting Groups}
\usage{
uniteInterestingGroups(object, ...)

\S4method{uniteInterestingGroups}{DataFrame}(object, interestingGroups)

\S4method{uniteInterestingGroups}{data.frame}(object, interestingGroups)

\S4method{uniteInterestingGroups}{tbl_df}(object, interestingGroups)
}
\arguments{
\item{object}{Object (e.g. \code{data.frame}) containing interesting groups
columns.}

\item{interestingGroups}{Character vector of interesting groups.}
}
\value{
Object of same class, containing \code{interestingGroups} column.
}
\description{
Create a single interesting groups column ("\code{interestingGroups}") used for
coloring in plots. When multiple interesting groups are present, unite into a
single column, delimited by "\code{:}".

Methods =====================================================================
}
\examples{
meta <- readSampleMetadataFile(
    "http://bcbiobase.seq.cloud/demultiplexed.csv"
)
meta <- uniteInterestingGroups(
    object = meta,
    interestingGroups = c("genotype", "sampleName")
)
meta[, "interestingGroups"]
}
\author{
Michael Steinbaugh
}
