% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{bcbio}
\alias{bcbio<-}
\alias{flatFiles}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{metrics}
\alias{plotDot}
\alias{plotGene}
\alias{plotQC}
\alias{plotViolin}
\alias{selectSamples}
\alias{tpm}
\title{S4 Generics}
\usage{
bcbio(object, ...)

bcbio(object, ...) <- value

flatFiles(object, ...)

interestingGroups(object, ...)

interestingGroups(object, ...) <- value

metrics(object, ...)

plotDot(object, ...)

plotGene(object, ...)

plotQC(object, ...)

plotViolin(object, ...)

selectSamples(object, ...)

tpm(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
Varies, depending upon the method.
}
\description{
S4 Generics
}
