% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertFormalInterestingGroups.R
\name{assertFormalInterestingGroups}
\alias{assertFormalInterestingGroups}
\title{Interesting Groups Formal Assert Check}
\usage{
assertFormalInterestingGroups(x, interestingGroups, severity = "stop")
}
\arguments{
\item{x}{Object supporting \code{\link[=colnames]{colnames()}}, typically a \code{data.frame}.}

\item{interestingGroups}{Interesting groups character vector.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Silent, stop on error.
}
\description{
Prevent unwanted downstream behavior when a missing interesting group
is requested by the user.
}
\examples{
data <- readSampleMetadataFile(
    "http://bcbiobase.seq.cloud/demultiplexed.csv"
)
assertFormalInterestingGroups(data, interestingGroups = "genotype")
}
\seealso{
Other Assert Check Functions: \code{\link{checkInterestingGroups}}
}
\author{
Michael Steinbaugh
}
