% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertFormalInterestingGroups.R
\name{assertFormalInterestingGroups}
\alias{assertFormalInterestingGroups}
\title{Interesting Groups Formal Assert Check}
\usage{
assertFormalInterestingGroups(x, interestingGroups, severity = "stop")
}
\arguments{
\item{x}{Object supporting \code{\link[=colnames]{colnames()}}, typically a \link{data.frame}.}

\item{interestingGroups}{Interesting groups character vector.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\description{
Prevent unwanted downstream behavior when a missing interesting group
is requested by the user.
}
\examples{
meta <- readSampleMetadataFile(paste(
    "http://bcbiobase.seq.cloud",
    "sample_metadata",
    "demultiplexed.xlsx",
    sep = "/"))

# Success
assertFormalInterestingGroups(meta, interestingGroups = "genotype")

# Failure
tryCatch(
    assertFormalInterestingGroups(meta, interestingGroups = "XXX"),
    error = function(e) e)
}
