% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertFormalInterestingGroups}
\alias{assertFormalInterestingGroups}
\title{Interesting Groups Formal Assert Check}
\usage{
assertFormalInterestingGroups(x, interestingGroups, severity = "stop")
}
\arguments{
\item{x}{Object supporting \code{\link[=colnames]{colnames()}}, typically a \code{data.frame}.}

\item{interestingGroups}{\code{character} vector defining the groups of interest.}

\item{severity}{How severe should the consequences of the assertion be?
Either "\code{stop}", "\code{warning}", "\code{message}", or "\code{none}".}
}
\value{
Silent, stop on error.
}
\description{
Prevent unwanted downstream behavior when a missing interesting group
is requested by the user.
}
\examples{
assertFormalInterestingGroups(
    x = sampleData(rse_bcb),
    interestingGroups = "treatment"
)
}
\seealso{
Other Assert Check Functions: \code{\link{assertFormalAnnotationCol}}
}
\author{
Michael Steinbaugh
}
