% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioBase-package.R
\docType{package}
\name{bcbioBase-package}
\alias{bcbioBase}
\alias{bcbioBase-package}
\title{bcbioBase}
\description{
Base functions and generics for bcbio R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://bioinformatics.sph.harvard.edu/bcbioBase}
  \item \url{https://github.com/hbc/bcbioBase}
  \item Report bugs at \url{https://github.com/hbc/bcbioBase/issues}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com}

Other contributors:
\itemize{
  \item Harvard Chan Bioinformatics Core \email{bioinformatics@hsph.harvard.edu} [copyright holder, funder]
}

}
