% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioBase-package.R
\docType{package}
\name{bcbioBase-package}
\alias{bcbioBase}
\alias{bcbioBase-package}
\title{bcbioBase}
\description{
Base functions and generics for bcbio R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://bioinformatics.sph.harvard.edu/bcbioBase}
  \item \url{https://github.com/hbc/bcbioBase}
  \item Report bugs at \url{https://github.com/hbc/bcbioBase/issues}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Authors:
\itemize{
  \item Lorena Pantano \email{lorena.pantano@gmail.com} (0000-0002-3859-3249)
  \item Rory Kirchner \email{roryk@alum.mit.edu} (0000-0003-4814-5885)
}

Other contributors:
\itemize{
  \item Harvard Chan Bioinformatics Core \email{bioinformatics@hsph.harvard.edu} [copyright holder, funder]
}

}
