% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{bcbioBaseTestsURL}
\alias{bcbioBaseTestsURL}
\title{bcbioBase test data URL}
\format{An object of class \code{character} of length 1.}
\usage{
bcbioBaseTestsURL
}
\description{
bcbioBase test data URL
}
\examples{
bcbioBaseTestsURL
}
\keyword{datasets}
