% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{commands}
\alias{commands}
\alias{getBarcodeCutoffFromCommands}
\alias{getLevelFromCommands}
\alias{getUMITypeFromCommands}
\title{Commands log parsing functions}
\usage{
getBarcodeCutoffFromCommands(log)

getLevelFromCommands(log)

getUMITypeFromCommands(log)
}
\arguments{
\item{log}{\code{character}.
Commands log.}
}
\value{
\code{atomic}.
}
\description{
Commands log parsing functions
}
\section{Functions}{
\itemize{
\item \code{getBarcodeCutoffFromCommands}: \code{integer(1)}.

\item \code{getLevelFromCommands}: \code{character(1)}. Return \code{"genes"} or \code{"transcripts"}.

\item \code{getUMITypeFromCommands}: \code{character(1)}.
}}

\note{
Updated 2022-03-07.
}
\examples{
file <- file.path(bcbioBaseTestsURL, "surecell-commands.log")
log <- import(file)
getBarcodeCutoffFromCommands(log)
getLevelFromCommands(log)
getUMITypeFromCommands(log)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
