% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{commands}
\alias{commands}
\alias{getBarcodeCutoffFromCommands}
\alias{getLevelFromCommands}
\alias{getUmiTypeFromCommands}
\title{Commands log parsing functions}
\usage{
getBarcodeCutoffFromCommands(log)

getLevelFromCommands(log)

getUmiTypeFromCommands(log)
}
\arguments{
\item{log}{\code{character}.
Commands log.}
}
\value{
\code{atomic}.
}
\description{
Commands log parsing functions
}
\section{Functions}{
\itemize{
\item \code{getBarcodeCutoffFromCommands()}: \code{integer(1)}.

\item \code{getLevelFromCommands()}: \code{character(1)}. Return \code{"genes"} or \code{"transcripts"}.

\item \code{getUmiTypeFromCommands()}: \code{character(1)}.

}}
\note{
Updated 2023-09-21.
}
\examples{
file <- file.path(bcbioBaseTestsUrl, "surecell-commands.log")
log <- import(file)
getBarcodeCutoffFromCommands(log)
getLevelFromCommands(log)
getUmiTypeFromCommands(log)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
