% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{commands}
\alias{commands}
\alias{getBarcodeCutoffFromCommands}
\title{Get from Commands Log}
\usage{
getBarcodeCutoffFromCommands(log)
}
\arguments{
\item{log}{\code{character}. Commands log.}
}
\value{
\code{integer}.
}
\description{
Get from Commands Log
}
\examples{
file <- file.path(bcbioBaseCacheURL, "bcbio_nextgen_commands.log")
log <- basejump::import(file)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
