% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToDropbox.R
\name{copyToDropbox}
\alias{copyToDropbox}
\title{Copy files to Dropbox}
\usage{
copyToDropbox(files, dir, rdsToken = NULL)
}
\arguments{
\item{files}{\code{character}.
Local file paths.}

\item{dir}{\code{character(1)}.
Relative path of remote Dropbox directory.}

\item{rdsToken}{\code{character(1)} or \code{NULL}.
RDS file token to use for Dropbox authentication.}
}
\value{
Invisible \code{list}.
rdrop2 output, including file paths.
}
\description{
Copy files to Dropbox
}
\note{
Requires the rdrop2 package to be installed.

Updated 2020-01-17.
}
\examples{
# > copyToDropbox(files = c("raw_counts.csv", "tpm.csv"), dir = "counts")
}
\author{
Michael Steinbaugh, Victor Barerra, John Hutchinson
}
