% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToDropbox.R
\name{copyToDropbox}
\alias{copyToDropbox}
\title{Copy Files to Dropbox}
\usage{
copyToDropbox(files, dir, rdsToken = NA)
}
\arguments{
\item{files}{Local file paths.}

\item{dir}{Relative path of remote Dropbox directory.}

\item{rdsToken}{RDS file token to use for Dropbox authentication.}
}
\value{
Invisibly return \link{list} of rdrop2 output.
}
\description{
Copy Files to Dropbox
}
\examples{
prepareTemplate("bibliography.bib")
dropboxDir <- file.path("bcbioBase_examples", "copyToDropbox")
suppressMessages(copyToDropbox(
    files = "bibliography.bib",
    dir = dropboxDir,
    rdsToken = system.file("extdata/token.rds", package = "bcbioBase")
))
rdrop2::drop_exists(file.path(dropboxDir, "bibliography.bib"))
unlink("bibliography.bib")
}
\author{
Michael Steinbaugh, Victor Barerra, John Hutchinson
}
