% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,SummarizedExperiment-method}
\title{Counts}
\usage{
\S4method{counts}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}.
}
\description{
Counts
}
\examples{
# SummarizedExperiment ====
x <- counts(rse_dds)
summary(x)
}
\seealso{
Other Data Functions: \code{\link{flatFiles}},
  \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleDirs}},
  \code{\link{sampleNames}},
  \code{\link{uniteInterestingGroups}}
}
