% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,SummarizedExperiment-method}
\title{Counts}
\usage{
\S4method{counts}{SummarizedExperiment}(object)
}
\description{
Counts
}
\examples{
x <- counts(rse_bcb)
glimpse(x)
}
\seealso{
Other Data Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{flatFiles}}, \code{\link{gene2symbol}},
  \code{\link{sampleDirs}}, \code{\link{selectSamples}},
  \code{\link{uniteInterestingGroups}}
}
