% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{readYAMLSampleData}
\alias{readYAMLSampleMetrics}
\title{Defunct Functions}
\usage{
readYAMLSampleData(...)

readYAMLSampleMetrics(...)
}
\value{
No value.
}
\description{
Defunct Functions
}
\keyword{internal}
