% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{bcbio_geom_abline}
\alias{bcbio_geom_label}
\alias{bcbio_geom_label_average}
\alias{bcbio_geom_label_repel}
\alias{readTx2gene}
\alias{readYAMLSampleData}
\alias{readYAMLSampleMetrics}
\title{Defunct functions}
\usage{
bcbio_geom_abline(...)

bcbio_geom_label(...)

bcbio_geom_label_average(...)

bcbio_geom_label_repel(...)

readTx2gene(...)

readYAMLSampleData(...)

readYAMLSampleMetrics(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{.Defunct}.
}
\description{
Defunct functions
}
\keyword{internal}
