% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{prepareSummarizedExperiment}
\title{Deprecated Functions}
\usage{
prepareSummarizedExperiment(...)
}
\value{
No value.
}
\description{
Deprecated Functions
}
\keyword{internal}
