% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{readLog}
\title{Deprecated functions}
\usage{
readLog(file)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\keyword{internal}
