% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{readLogFile}
\alias{readSampleMetadataFile}
\alias{sampleMetadata}
\alias{sampleMetadata<-}
\alias{flatFiles}
\alias{metrics}
\alias{plotCorrelationHeatmap}
\alias{plotGene}
\alias{plotHeatmap}
\alias{plotQC}
\alias{plotQuantileHeatmap}
\alias{prepareSummarizedExperiment}
\alias{bcbio_geom_abline}
\alias{bcbio_geom_label}
\alias{bcbio_geom_label_average}
\alias{bcbio_geom_label_repel}
\alias{readLog}
\alias{readTx2gene}
\title{Deprecated Functions}
\usage{
readLogFile(...)

readSampleMetadataFile(...)

sampleMetadata(object, ...)

sampleMetadata(object) <- value

flatFiles(object, ...)

metrics(object, ...)

plotCorrelationHeatmap(object, ...)

plotGene(object, ...)

plotHeatmap(object, ...)

plotQC(object, ...)

plotQuantileHeatmap(object, ...)

prepareSummarizedExperiment(...)

bcbio_geom_abline(...)

bcbio_geom_label(...)

bcbio_geom_label_average(...)

bcbio_geom_label_repel(...)

readLog(file)

readTx2gene(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{.Deprecated()}.
}
\description{
Deprecated Functions
}
\keyword{internal}
