% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{checkInterestingGroups}
\alias{prepareSampleMetadata}
\alias{readLogFile}
\alias{readSampleMetadataFile}
\alias{sampleMetadata}
\alias{sampleMetadata<-}
\alias{sampleYAML}
\alias{sampleYAMLMetadata}
\alias{sampleYAMLMetrics}
\title{Defunct or Deprecated Functions}
\usage{
checkInterestingGroups(...)

prepareSampleMetadata(...)

readLogFile(...)

readSampleMetadataFile(...)

sampleMetadata(object, ...)

sampleMetadata(object) <- value

sampleYAML(...)

sampleYAMLMetadata(...)

sampleYAMLMetrics(...)
}
\value{
No value.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
