% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{checkInterestingGroups}
\alias{prepareSampleMetadata}
\alias{readLogFile}
\alias{readSampleMetadataFile}
\title{Defunct or Deprecated Functions}
\usage{
checkInterestingGroups(...)

prepareSampleMetadata(...)

readLogFile(...)

readSampleMetadataFile(...)
}
\value{
No value.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
