% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{readLogFile}
\alias{readSampleMetadataFile}
\alias{sampleMetadata}
\alias{sampleMetadata<-}
\alias{prepareSummarizedExperiment}
\title{Deprecated Functions}
\usage{
readLogFile(...)

readSampleMetadataFile(...)

sampleMetadata(object, ...)

sampleMetadata(object) <- value

prepareSummarizedExperiment(...)
}
\value{
No value.
}
\description{
Deprecated Functions
}
\keyword{internal}
