% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{readLogFile}
\alias{readSampleMetadataFile}
\alias{sampleMetadata}
\alias{sampleMetadata<-}
\alias{flatFiles}
\alias{metrics}
\alias{plotCorrelationHeatmap}
\alias{plotGene}
\alias{plotHeatmap}
\alias{plotQC}
\alias{plotQuantileHeatmap}
\alias{prepareSummarizedExperiment}
\title{Deprecated Functions}
\usage{
readLogFile(...)

readSampleMetadataFile(...)

sampleMetadata(object, ...)

sampleMetadata(object) <- value

flatFiles(object, ...)

metrics(object, ...)

plotCorrelationHeatmap(object, ...)

plotGene(object, ...)

plotHeatmap(object, ...)

plotQC(object, ...)

plotQuantileHeatmap(object, ...)

prepareSummarizedExperiment(...)
}
\value{
No value.
}
\description{
Deprecated Functions
}
\keyword{internal}
