% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{bcbio_geom_abline}
\alias{bcbio_geom_label}
\alias{bcbio_geom_label_average}
\alias{bcbio_geom_label_repel}
\alias{readLog}
\alias{readTx2gene}
\title{Deprecated functions}
\usage{
bcbio_geom_abline(...)

bcbio_geom_label(...)

bcbio_geom_label_average(...)

bcbio_geom_label_repel(...)

readLog(file)

readTx2gene(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\keyword{internal}
