% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{assertFormalInterestingGroups}
\alias{convertGenesToSymbols}
\alias{gene2symbol}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleNames}
\alias{sanitizeSampleData}
\alias{selectSamples}
\alias{uniteInterestingGroups}
\alias{separatorBar}
\alias{updateMessage}
\alias{flatFiles}
\alias{metrics}
\alias{plotCorrelationHeatmap}
\alias{plotGene}
\alias{plotHeatmap}
\alias{plotQC}
\alias{plotQuantileHeatmap}
\alias{prepareSummarizedExperiment}
\alias{prepareTemplate}
\alias{bcbio_geom_abline}
\alias{bcbio_geom_label}
\alias{bcbio_geom_label_average}
\alias{bcbio_geom_label_repel}
\title{Deprecated Functions}
\usage{
flatFiles(object, ...)

metrics(object, ...)

plotCorrelationHeatmap(object, ...)

plotGene(object, ...)

plotHeatmap(object, ...)

plotQC(object, ...)

plotQuantileHeatmap(object, ...)

prepareSummarizedExperiment(...)

bcbio_geom_abline(...)

bcbio_geom_label(...)

bcbio_geom_label_average(...)

bcbio_geom_label_repel(...)
}
\value{
No value.
}
\description{
Deprecated Functions
}
\keyword{internal}
