% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{readLogFile}
\alias{readSampleMetadataFile}
\alias{sampleMetadata}
\alias{sampleMetadata<-}
\alias{prepareSummarizedExperiment}
\alias{bcbio_geom_abline}
\alias{bcbio_geom_label}
\alias{bcbio_geom_label_average}
\alias{bcbio_geom_label_repel}
\alias{readLog}
\alias{readTx2gene}
\title{Deprecated Functions}
\usage{
readLogFile(...)

readSampleMetadataFile(...)

sampleMetadata(object, ...)

sampleMetadata(object) <- value

prepareSummarizedExperiment(...)

bcbio_geom_abline(...)

bcbio_geom_label(...)

bcbio_geom_label_average(...)

bcbio_geom_label_repel(...)

readLog(file)

readTx2gene(...)
}
\value{
No value.
}
\description{
Deprecated Functions
}
\keyword{internal}
