% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecated.R,
%   R/methods-flatFiles.R
\docType{methods}
\name{flatFiles}
\alias{flatFiles}
\alias{deprecated}
\alias{checkInterestingGroups}
\alias{prepareSampleMetadata}
\alias{readLogFile}
\alias{readSampleMetadataFile}
\alias{flatFiles,SummarizedExperiment-method}
\title{Defunct or Deprecated Functions}
\usage{
flatFiles(object, ...)

checkInterestingGroups(...)

prepareSampleMetadata(...)

readLogFile(...)

readSampleMetadataFile(...)

\S4method{flatFiles}{SummarizedExperiment}(object)
}
\value{
No value.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
