% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{sampleMetadata}
\alias{sampleMetadata<-}
\alias{flatFiles}
\title{Defunct or Deprecated Functions}
\usage{
sampleMetadata(object, ...)

sampleMetadata(object, ...) <- value

flatFiles(object)
}
\value{
No value.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
