% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-flatFIles.R,
%   R/deprecated-sampleMetadata.R, R/deprecated.R
\docType{methods}
\name{flatFiles}
\alias{flatFiles}
\alias{flatFiles,SummarizedExperiment-method}
\alias{sampleMetadata}
\alias{sampleMetadata<-}
\alias{deprecated}
\title{Defunct or Deprecated Functions}
\usage{
flatFiles(object, ...)

\S4method{flatFiles}{SummarizedExperiment}(object)

sampleMetadata(object, ...)

sampleMetadata(object, ...) <- value
}
\value{
No value.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
