% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecated-sampleMetadata.R,
%   R/deprecated.R, R/methods-flatFiles.R
\docType{methods}
\name{flatFiles}
\alias{flatFiles}
\alias{sampleMetadata}
\alias{sampleMetadata<-}
\alias{sampleMetadata,SummarizedExperiment-method}
\alias{sampleMetadata<-,SummarizedExperiment-method}
\alias{deprecated}
\alias{prepareSampleMetadata}
\alias{readLogFile}
\alias{readSampleMetadataFile}
\alias{flatFiles,SummarizedExperiment-method}
\title{Defunct or Deprecated Functions}
\usage{
flatFiles(object, ...)

sampleMetadata(object, ...)

sampleMetadata(object, ...) <- value

\S4method{sampleMetadata}{SummarizedExperiment}(object, ...)

\S4method{sampleMetadata}{SummarizedExperiment}(object, ...) <- value

prepareSampleMetadata(...)

readLogFile(...)

readSampleMetadataFile(...)

\S4method{flatFiles}{SummarizedExperiment}(object)
}
\value{
No value.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
