% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-featureData.R
\docType{methods}
\name{featureData}
\alias{featureData}
\alias{featureData,ANY-method}
\title{Feature Data}
\usage{
\S4method{featureData}{ANY}(object, return = c("DataFrame", "data.frame",
  "AsIs"), ...)
}
\arguments{
\item{object}{Object.}

\item{return}{\code{DataFrame}, \code{data.frame}, or \code{AsIs}.}

\item{...}{Additional arguments.}
}
\value{
Gene/transcript (feature) metadata. Note that the features are
defined in the rows, similar to \code{\link[=rowData]{rowData()}}.
}
\description{
Return the gene or transcript (feature) metadata.
}
\note{
This is a complement to the standard \code{\link[=rowData]{rowData()}} function.
}
\seealso{
Other Metadata Functions: \code{\link{featureData<-,ANY-method}}
}
