% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/flatFiles-methods.R
\docType{methods}
\name{flatFiles}
\alias{flatFiles}
\alias{flatFiles}
\alias{flatFiles,SummarizedExperiment-method}
\title{Flat Files from S4 Object}
\usage{
flatFiles(object, ...)

\S4method{flatFiles}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list} containing the slots of an S4 object.
}
\description{
Extract the slots inside an S4 object for archival storage.
}
\examples{
# SummarizedExperiment ====
x <- flatFiles(rse_dds)
class(x)
names(x)
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleDirs}},
  \code{\link{sampleNames}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
