% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-flatFiles.R
\name{flatFiles}
\alias{flatFiles}
\title{Flat Files from S4 Object}
\value{
\code{list} containing the slots of an S4 object.
}
\description{
Extract the slots inside an S4 object for archival storage.
}
\seealso{
Other Data Functions: \code{\link{gene2symbol}},
  \code{\link{sampleDirs}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
