% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,SummarizedExperiment-method}
\title{Gene to Symbol Mappings}
\usage{
gene2symbol(object)

\S4method{gene2symbol}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{data.frame} containing gene identifier and gene name (aka symbol)
mappings.
}
\description{
Gene to Symbol Mappings
}
\seealso{
Other Data Functions: \code{\link{flatFiles}},
  \code{\link{sampleDirs}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
