% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,SummarizedExperiment-method}
\title{Gene to Symbol Mappings}
\usage{
\S4method{gene2symbol}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{data.frame} containing Ensembl gene identifier and gene name
(aka symbol) mappings.
}
\description{
Gene to Symbol Mappings
}
\author{
Michael Steinbaugh
}
