% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{general}
\alias{general}
\title{General Arguments}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{x}{Object.}

\item{...}{Additional arguments.}

\item{file}{\code{string}. File path.}

\item{uploadDir}{\code{string}. File path to bcbio run upload directory.}
}
\value{
No value.
}
\description{
General Arguments
}
\keyword{internal}
