% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromCommandsLog.R
\name{getBarcodeCutoffFromCommandsLog}
\alias{getBarcodeCutoffFromCommandsLog}
\title{Get Cellular Barcode Cutoff from Commands Log}
\usage{
getBarcodeCutoffFromCommandsLog(log)
}
\arguments{
\item{log}{\code{character}. Commands log.}
}
\value{
\code{integer}.
}
\description{
Get Cellular Barcode Cutoff from Commands Log
}
\seealso{
Other Data Functions: \code{\link{getLevelFromCommandsLog}},
  \code{\link{getUMITypeFromCommandsLog}},
  \code{\link{projectDir}}, \code{\link{runDate}},
  \code{\link{sampleDirs}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Data Functions}
