% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromYAML.R
\name{getGTFFileFromYAML}
\alias{getGTFFileFromYAML}
\title{Get GTF File from YAML}
\usage{
getGTFFileFromYAML(yaml)
}
\arguments{
\item{yaml}{\code{list}. Project summary YAML.}
}
\value{
\code{string}. GTF file path.
}
\description{
Get GTF File from YAML
}
\examples{
\dontrun{
getGTFFileFromYAML(yaml)
}
}
