% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGTFFileFromYAML.R
\name{getGTFFileFromYAML}
\alias{getGTFFileFromYAML}
\title{Get GTF File from YAML}
\usage{
getGTFFileFromYAML(yaml)
}
\value{
\code{string}.
}
\description{
Get GTF File from YAML
}
\examples{
file <- file.path(bcbioBaseCacheURL, "summary.yaml")
yaml <- import(file)
x <- getGTFFileFromYAML(yaml)
print(x)
}
\author{
Michael Steinbaugh
}
