% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGTFFileFromYAML.R
\name{getGTFFileFromYAML}
\alias{getGTFFileFromYAML}
\title{Get GTF file path from YAML}
\usage{
getGTFFileFromYAML(yaml)
}
\arguments{
\item{yaml}{\code{list}.
YAML metadata.}
}
\value{
\code{character(1)} or \code{NULL}.
File path if the file exists. \code{NULL} if the file does not exist.
}
\description{
Get GTF file path from YAML
}
\note{
Updated 2020-01-17.
}
\examples{
file <- file.path(bcbioBaseTestsURL, "summary.yaml")
yaml <- basejump::import(file)
x <- getGTFFileFromYAML(yaml)
print(x)
}
\author{
Michael Steinbaugh
}
