% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromCommandsLog.R
\name{getLevelFromCommandsLog}
\alias{getLevelFromCommandsLog}
\title{Get Gene or Transcript Level from Commands Log}
\usage{
getLevelFromCommandsLog(log)
}
\arguments{
\item{log}{\code{character}. Commands log.}
}
\description{
Get Gene or Transcript Level from Commands Log
}
\author{
Michael Steinbaugh, Rory Kirchner
}
