% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleDataFromYaml.R
\name{getSampleDataFromYaml}
\alias{getSampleDataFromYaml}
\title{Get sample data from YAML}
\usage{
getSampleDataFromYaml(yaml)
}
\arguments{
\item{yaml}{\code{list}.
YAML metadata.}
}
\value{
\code{DataFrame}.
}
\description{
Get sample data from YAML
}
\note{
Updated 2022-03-07.
}
\examples{
file <- file.path(bcbioBaseTestsUrl, "summary.yaml")
yaml <- import(file)
x <- getSampleDataFromYaml(yaml)
summary(x)
colnames(x)
}
\author{
Michael Steinbaugh
}
