% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleDataFromYAML.R
\name{getSampleDataFromYAML}
\alias{getSampleDataFromYAML}
\title{Get sample data from YAML}
\usage{
getSampleDataFromYAML(yaml)
}
\arguments{
\item{yaml}{\code{list}.
Project summary YAML.}
}
\value{
\code{DataFrame}.
}
\description{
Get sample data from YAML
}
\examples{
file <- file.path(bcbioBaseCacheURL, "summary.yaml")
yaml <- basejump::import(file)
x <- getSampleDataFromYAML(yaml)
summary(x)
colnames(x)
}
\author{
Michael Steinbaugh
}
