% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromCommandsLog.R
\name{getUMITypeFromCommandsLog}
\alias{getUMITypeFromCommandsLog}
\title{Get UMI Type from Commands Log}
\usage{
getUMITypeFromCommandsLog(log)
}
\arguments{
\item{log}{\code{character}. Commands log.}
}
\description{
Get UMI Type from Commands Log
}
\author{
Rory Kirchner, Michael Steinbaugh
}
