% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importProgramVersions.R
\name{importProgramVersions}
\alias{importProgramVersions}
\title{Import program versions}
\usage{
importProgramVersions(file)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{DataFrame}.
}
\description{
Import program versions
}
\note{
bcbio doesn't save program versions when run in fast mode.

Updated 2019-08-27.
}
\examples{
file <- file.path(bcbioBaseTestsURL, "programs.txt")
x <- importProgramVersions(file)
print(x)
}
\author{
Michael Steinbaugh
}
