% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{metadataPriorityCols}
\alias{metadataPriorityCols}
\title{Metadata Priority Columns}
\format{An object of class \code{character} of length 2.}
\usage{
metadataPriorityCols
}
\description{
Metadata Priority Columns
}
\examples{
metadataPriorityCols
}
\keyword{internal}
