% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimalSampleData.R
\name{minimalSampleData}
\alias{minimalSampleData}
\title{Minimal Sample Data}
\usage{
minimalSampleData(sample)
}
\arguments{
\item{sample}{\code{character}. Sample names (e.g. "description" in bcbio YAML).}
}
\value{
\code{DataFrame}.
}
\description{
Minimal Sample Data
}
\examples{
x <- minimalSampleData(sample = c("sample 1", "sample 2"))
print(x)
}
\seealso{
Other Data Functions: \code{\link{projectDir}},
  \code{\link{sampleDirs}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
