% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{projectDir}{\code{string}. Project directory path.}

\item{uploadDir}{\code{string}. File path to bcbio run upload directory.}
}
\value{
No value.
}
\description{
Parameters
}
\keyword{internal}
