% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{projectDir}{\code{character(1)}. Project directory path.}

\item{uploadDir}{\code{character(1)}. Path to final upload directory. This path is
set when running "\code{bcbio_nextgen -w template}".}

\item{yaml}{\code{list}. Project summary YAML.}
}
\value{
No value.
}
\description{
Parameters
}
\keyword{internal}
