% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotCorrelationHeatmap-methods.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,SummarizedExperiment-method}
\title{Plot Correlation Heatmap}
\usage{
plotCorrelationHeatmap(object, ...)

\S4method{plotCorrelationHeatmap}{SummarizedExperiment}(object,
  interestingGroups, method = c("pearson", "spearman"),
  clusteringMethod = "ward.D2", showRownames = TRUE, showColnames = TRUE,
  treeheightRow = 0L, treeheightCol = 50L, color = viridis,
  legendColor = NULL, borderColor = NULL, title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments should be formatted in camel case, not snake case.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{method}{Correlation coefficient (or covariance) method to be computed.
Defaults to "\code{pearson}" but "\code{spearman}" can also be used. Consult the
\code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{clusteringMethod}{Clustering method. Accepts the same values as
\code{\link[stats:hclust]{stats::hclust()}}.}

\item{showRownames}{Show row or column names.}

\item{showColnames}{Show row or column names.}

\item{treeheightRow}{Size of the row and column dendrograms.
Use "\code{0}" to disable.}

\item{treeheightCol}{Size of the row and column dendrograms.
Use "\code{0}" to disable.}

\item{color}{Hexadecimal color function to use for plot. We recommend any of
these from the viridis package:
\itemize{
\item \code{\link[viridis:viridis]{viridis::viridis()}} (\emph{default}).
\item \code{\link[viridis:inferno]{viridis::inferno()}}.
\item \code{\link[viridis:magma]{viridis::magma()}}.
\item \code{\link[viridis:plasma]{viridis::plasma()}}.
}}

\item{legendColor}{Hexadecimal color function to use for legend labels.}

\item{borderColor}{\emph{Optional.} Border color. Disabled by default for
improved aesthetics.}

\item{title}{\emph{Optional.} Plot title.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
Construct a correlation heatmap comparing the columns of the matrix.
}
\examples{
# SummarizedExperiment ====
plotCorrelationHeatmap(rse_dds, interestingGroups = "condition")
plotCorrelationHeatmap(rse_dds, color = magma, legendColor = viridis)
plotCorrelationHeatmap(rse_dds, color = NULL)
}
\seealso{
\itemize{
\item \code{\link[stats:cor]{stats::cor()}}.
\item \code{\link[stats:hclust]{stats::hclust()}}.
\item \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}

Other Plot Functions: \code{\link{plotHeatmap}},
  \code{\link{plotQuantileHeatmap}}
}
\author{
Michael Steinbaugh
}
