% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCorrelationHeatmap.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,matrix-method}
\alias{plotCorrelationHeatmap,dgCMatrix-method}
\alias{plotCorrelationHeatmap,dgTMatrix-method}
\alias{plotCorrelationHeatmap,SummarizedExperiment-method}
\title{Plot Correlation Heatmap}
\usage{
plotCorrelationHeatmap(object, ...)

\S4method{plotCorrelationHeatmap}{matrix}(object, method = c("pearson",
  "spearman"), clusteringMethod = "ward.D2", annotationCol = NULL,
  color = viridis, legendColor = NULL, borderColor = NULL, title = TRUE,
  ...)

\S4method{plotCorrelationHeatmap}{dgCMatrix}(object, method = c("pearson",
  "spearman"), clusteringMethod = "ward.D2", annotationCol = NULL,
  color = viridis, legendColor = NULL, borderColor = NULL, title = TRUE,
  ...)

\S4method{plotCorrelationHeatmap}{dgTMatrix}(object, method = c("pearson",
  "spearman"), clusteringMethod = "ward.D2", annotationCol = NULL,
  color = viridis, legendColor = NULL, borderColor = NULL, title = TRUE,
  ...)

\S4method{plotCorrelationHeatmap}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments should be formatted in camel case, not snake case.}

\item{method}{Correlation coefficient (or covariance) method to be computed.
Defaults to "\code{pearson}" but "\code{spearman}" can also be used. Consult the
\code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{clusteringMethod}{Clustering method. Accepts the same values as
\code{\link[stats:hclust]{stats::hclust()}}.}

\item{annotationCol}{\emph{Optional.} \code{data.frame} that defines annotation
mappings for the columns.}

\item{color}{Colors to use for plot. Defaults to the \code{\link[viridis:viridis]{viridis::viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels.}

\item{borderColor}{\emph{Optional.} Border color. Disabled by default for
improved aesthetics.}

\item{title}{\emph{Optional.} Plot title.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
Construct a correlation heatmap comparing the columns of the matrix.
}
\examples{
# SummarizedExperiment ====
plotCorrelationHeatmap(rse_dds)

# matrix ====
mat <- assay(rse_dds)
plotCorrelationHeatmap(mat)
}
\seealso{
\itemize{
\item \code{\link[stats:cor]{stats::cor()}}.
\item \code{\link[stats:hclust]{stats::hclust()}}.
\item \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}

Other Plot Heatmap Functions: \code{\link{plotHeatmap}},
  \code{\link{plotQuantileHeatmap}}
}
\author{
Michael Steinbaugh
}
