% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotHeatmap.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,matrix-method}
\alias{plotHeatmap,dgCMatrix-method}
\alias{plotHeatmap,dgTMatrix-method}
\alias{plotHeatmap,SummarizedExperiment-method}
\title{Plot Heatmap}
\usage{
plotHeatmap(object, ...)

\S4method{plotHeatmap}{matrix}(object, scale = c("row", "column", "none"),
  annotationCol = NULL, clusterRows = TRUE, clusterCols = TRUE,
  showRownames = FALSE, showColnames = TRUE, treeheightRow = 0L,
  treeheightCol = 50L, color = viridis, legendColor = NULL,
  borderColor = NULL, title = NULL, ...)

\S4method{plotHeatmap}{dgCMatrix}(object, scale = c("row", "column", "none"),
  annotationCol = NULL, clusterRows = TRUE, clusterCols = TRUE,
  showRownames = FALSE, showColnames = TRUE, treeheightRow = 0L,
  treeheightCol = 50L, color = viridis, legendColor = NULL,
  borderColor = NULL, title = NULL, ...)

\S4method{plotHeatmap}{dgTMatrix}(object, scale = c("row", "column", "none"),
  annotationCol = NULL, clusterRows = TRUE, clusterCols = TRUE,
  showRownames = FALSE, showColnames = TRUE, treeheightRow = 0L,
  treeheightCol = 50L, color = viridis, legendColor = NULL,
  borderColor = NULL, title = NULL, ...)

\S4method{plotHeatmap}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments should be formatted in camel case, not snake case.}

\item{scale}{\code{character} indicating whether the values should be centered and
scaled in either the row or column direction ("\code{row}", "\code{column}"), or
remain unscaled ("\code{none}").}

\item{annotationCol}{\emph{Optional.} \code{data.frame} that defines annotation
mappings for the columns.}

\item{clusterRows, clusterCols}{\code{logical} determining if rows or columns
should be arranged with hierarchical clustering. Alternatively, can define
an \code{hclust} object.}

\item{showRownames, showColnames}{Show row or column names.}

\item{treeheightRow, treeheightCol}{Size of the row and column dendrograms.
Use "\code{0}" to disable.}

\item{color}{Colors to use for plot. Defaults to the \code{\link[viridis:viridis]{viridis::viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels.}

\item{borderColor}{\emph{Optional.} Border color. Disabled by default for
improved aesthetics.}

\item{title}{\emph{Optional.} Plot title.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
Construct a simple heatmap. By default, row- and column-wise hierarchical
clustering is performed using the Ward method, but this behavior can be
overrided by setting \code{clusterRows} or \code{clusterCols} to \code{FALSE}.
}
\examples{
# SummarizedExperiment ====
plotHeatmap(rse_dds)

# Disable column clustering
plotHeatmap(rse_dds, clusterCols = FALSE)
}
\seealso{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.

Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotQuantileHeatmap}}
}
