% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotQuantileHeatmap.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,matrix-method}
\alias{plotQuantileHeatmap,dgCMatrix-method}
\alias{plotQuantileHeatmap,dgTMatrix-method}
\alias{plotQuantileHeatmap,SummarizedExperiment-method}
\title{Plot Heatmap with Quantile Breaks}
\usage{
plotQuantileHeatmap(object, ...)

\S4method{plotQuantileHeatmap}{matrix}(object, n = 10L,
  annotationCol = NULL, clusterCols = TRUE, clusterRows = TRUE,
  showColnames = TRUE, showRownames = FALSE, color = viridis,
  legendColor = inferno, borderColor = NULL, title = NULL, ...)

\S4method{plotQuantileHeatmap}{dgCMatrix}(object, n = 10L,
  annotationCol = NULL, clusterCols = TRUE, clusterRows = TRUE,
  showColnames = TRUE, showRownames = FALSE, color = viridis,
  legendColor = inferno, borderColor = NULL, title = NULL, ...)

\S4method{plotQuantileHeatmap}{dgTMatrix}(object, n = 10L,
  annotationCol = NULL, clusterCols = TRUE, clusterRows = TRUE,
  showColnames = TRUE, showRownames = FALSE, color = viridis,
  legendColor = inferno, borderColor = NULL, title = NULL, ...)

\S4method{plotQuantileHeatmap}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments must be formatted in camel case, not snake case.}

\item{n}{The number of breaks to create.}

\item{annotationCol}{\emph{Optional.} \code{data.frame} that defines annotation
mappings for the columns.}

\item{clusterCols}{Logical determining if columns should be arranged with
hierarchical clustering. Alternatively, can define an \code{hclust} object.}

\item{clusterRows}{Logical determining if rows should be arranged with
hierarchical clustering. Alternatively, can define an \code{hclust} object.}

\item{showColnames}{Show column names.}

\item{showRownames}{Show row names.}

\item{color}{Colors to use for plot. Defaults to the \code{\link[=viridis]{viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[=viridis]{viridis()}} palette.}

\item{borderColor}{Border color.}

\item{title}{\emph{Optional.} Plot title.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
Plot Heatmap with Quantile Breaks
}
\examples{
# SummarizedExperiment ====
plotQuantileHeatmap(rse_small)

# matrix ====
mat <- assay(rse_small)
plotQuantileHeatmap(mat)
}
\seealso{
Other Plot Heatmap Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotHeatmap}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
