% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSampleData.R
\name{prepareSampleData}
\alias{prepareSampleData}
\title{Prepare Sample Data}
\usage{
prepareSampleData(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{data.frame}.
}
\description{
Prepare Sample Data
}
\seealso{
Other Prepare Functions: \code{\link{prepareSummarizedExperiment}},
  \code{\link{prepareTemplate}}
}
\author{
Michael Steinbaugh
}
