% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSummarizedExperiment.R
\name{prepareSummarizedExperiment}
\alias{prepareSummarizedExperiment}
\title{Prepare Summarized Experiment}
\usage{
prepareSummarizedExperiment(assays, rowRanges, colData, metadata = NULL,
  isSpike = NULL)
}
\arguments{
\item{assays}{List containing RNA-seq count matrices with matching
dimensions. Counts can be passed in either dense (\code{matrix}) or sparse
(\code{dgCMatrix}, \code{dgTMatrix}) format.}

\item{rowRanges}{\code{GRanges} describing assay rows. Must contain genomic
ranges. Can be left \code{NULL} if the genome is poorly annotated and/or ranges
aren't available from AnnotationHub.}

\item{colData}{\code{DataFrame} \code{data.frame}, or \code{matrix} describing assay
columns.}

\item{metadata}{\emph{Optional}. Metadata \code{list}.}

\item{isSpike}{Character vector of spike-in sequence names
(matches against the rownames of the object).}
}
\value{
\code{RangedSummarizedExperiment}.
}
\description{
This is a utility wrapper for \code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment()}}
that provides automatic subsetting for row and column data. It also provides
automatic handling of FASTA spike-ins.
}
\details{
This function also provides automatic metadata slotting of multiple useful
environment parameters:
\itemize{
\item \code{date}: Today's date, returned from \code{\link[base:Sys.Date]{base::Sys.Date()}}.
\item \code{wd}: Working directory, returned from \code{\link[base:getwd]{base::getwd()}}.
\item \code{utilsSessionInfo}: \code{\link[utils:sessionInfo]{utils::sessionInfo()}}.
\item \code{devtoolsSessionInfo}: \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}}.
}
}
\examples{
# Example matrix
genes <- c(
    "EGFP",  # spike
    "gene1",
    "gene2",
    "gene3"
)
samples <- c(
    "sample1",
    "sample2",
    "sample3",
    "sample4"
)
mat <- matrix(
    seq(1L:16L),
    nrow = 4L,
    ncol = 4L,
    dimnames = list(genes, samples)
)

# Primary assay must be named "counts"
assays = list("counts" = mat)

# Leave out the unannotated EGFP spike-in
rowRanges <- GRanges(
    seqnames = c("1", "1", "1"),
    ranges = IRanges(
        start = c(1L, 101L, 201L),
        end = c(100L, 200L, 300L)
    )
)
names(rowRanges) <- c("gene1", "gene2", "gene3")

colData <- data.frame(
    "genotype" = c(
        "wildtype",
        "wildtype",
        "knockout",
        "knockout"
    ),
    "age" = c(3L, 6L, 3L, 6L),
    row.names = samples
)

prepareSummarizedExperiment(
    assays = assays,
    rowRanges = rowRanges,
    colData = colData,
    isSpike = "EGFP"
)
}
\author{
Michael Steinbaugh
}
