% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-prepareTemplate.R
\docType{methods}
\name{prepareTemplate}
\alias{prepareTemplate}
\alias{prepareTemplate}
\alias{prepareTemplate,missing-method}
\alias{prepareTemplate,character-method}
\title{Prepare R Markdown Template File}
\usage{
prepareTemplate(object, ...)

\S4method{prepareTemplate}{missing}(object, sourceDir = NULL)

\S4method{prepareTemplate}{character}(object, sourceDir = NULL)
}
\arguments{
\item{object}{\emph{Optional}. File name. If \code{NULL} (default), download the
default dependency files for a new experiment.}

\item{...}{Additional arguments.}

\item{sourceDir}{Source directory, typically a URL, where the dependency
files are located.}
}
\value{
No value.
}
\description{
If the required template dependency files aren't present, download latest
versions from the package website. \emph{Existing files are never overwritten.}
}
\details{
By default, this function will create local copies of these files:
\itemize{
\item \code{_footer.Rmd}
\item \code{_header.Rmd}
\item \code{_output.yaml}
\item \code{bibliography.bib}
\item \code{setup.R}
}
}
\examples{
# Copy all of the default shared template files
prepareTemplate()
unlink(c(
    "_footer.Rmd",
    "_header.Rmd",
    "_output.yaml",
    "bibliography.bib",
    "setup.R"
))

# Request individual files
prepareTemplate("setup.R")
unlink("setup.R")

# Load the shared files from bcbioSingleCell
\dontrun{
prepareTemplate(
    sourceDir = system.file(
        "rmarkdown/shared",
        package = "bcbioSingleCell")
)
}

}
