% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectDir.R
\name{projectDir}
\alias{projectDir}
\title{Project directory}
\usage{
projectDir(uploadDir)
}
\arguments{
\item{uploadDir}{\code{character(1)}.
Final upload directory path.}
}
\value{
\code{character(1)}.
Dated project directory (e.g. "2018-01-01_rnaseq").
}
\description{
Project directory
}
\note{
This will pick the latest dated directory and warn the user, if bcbio
has been run multiple times to the same upload directory.

Updated 2019-08-05.
}
\examples{
uploadDir <- system.file("extdata/bcbio", package = "bcbioBase")
x <- projectDir(uploadDir)
basename(x)
}
\author{
Michael Steinbaugh
}
