% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectDir.R
\name{projectDir}
\alias{projectDir}
\title{Project Directory}
\usage{
projectDir(uploadDir)
}
\arguments{
\item{uploadDir}{\code{string}. File path to bcbio run upload directory.}
}
\value{
\code{string}. Dated project directory (e.g. "2018-01-01_rnaseq").
}
\description{
Project Directory
}
\note{
This will pick the latest dated directory and warn the user, if bcbio
has been run multiple times to the same upload directory.
}
\examples{
uploadDir <- system.file("extdata/bcbio", package = "bcbioBase")
x <- projectDir(uploadDir)
basename(x)
}
\seealso{
Other Data Functions: \code{\link{runDate}},
  \code{\link{sampleDirs}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
