% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataVersions.R
\name{readDataVersions}
\alias{readDataVersions}
\title{Read Data Versions}
\usage{
readDataVersions(file)
}
\arguments{
\item{file}{Data versions CSV file.}
}
\value{
\code{data.frame}.
}
\description{
Read Data Versions
}
\note{
bcbio doesn't save data versions when run in fast mode.
}
\examples{
url <- paste(
    "http://bcbiobase.seq.cloud",
    "bcbio",
    "data_versions.csv",
    sep = "/"
)
readDataVersions(url) \%>\% glimpse()
}
\seealso{
Other Read Functions: \code{\link{readLogFile}},
  \code{\link{readProgramVersions}},
  \code{\link{readSampleMetadataFile}}
}
