% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataVersions.R
\name{readDataVersions}
\alias{readDataVersions}
\title{Read Data Versions}
\usage{
readDataVersions(file, quiet = FALSE)
}
\arguments{
\item{file}{Data versions CSV file.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
\link{data.frame}.
}
\description{
Read Data Versions
}
\examples{
url <- file.path(
    "http://bcbiobase.seq.cloud",
    "bcbio",
    "data_versions.csv")
readDataVersions(url) \%>\% glimpse()
}
\seealso{
Other File Utilities: \code{\link{readLogFile}}
}
