% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataVersions.R
\name{readDataVersions}
\alias{readDataVersions}
\title{Read Data Versions}
\usage{
readDataVersions(file)
}
\arguments{
\item{file}{Data versions CSV file.}
}
\value{
\code{data.frame}.
}
\description{
Read Data Versions
}
\note{
bcbio doesn't save data versions when run in fast mode.
}
\examples{
readDataVersions(
    "http://bcbiobase.seq.cloud/data_versions.csv"
) \%>\%
    glimpse()
}
\seealso{
Other Read Functions: \code{\link{readLogFile}},
  \code{\link{readProgramVersions}},
  \code{\link{readSampleMetadataFile}}
}
