% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataVersions.R
\name{readDataVersions}
\alias{readDataVersions}
\title{Read data versions}
\usage{
readDataVersions(file)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{DataFrame}.
}
\description{
Read data versions
}
\note{
The \code{data_versions.csv} file is only generated for special genomes
containing additional information (e.g. the built-in "\code{hg38}" build).
}
\examples{
file <- file.path(bcbioBaseCacheURL, "data_versions.csv")
x <- readDataVersions(file)
print(x)
}
\author{
Michael Steinbaugh
}
