% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataVersions.R
\name{readDataVersions}
\alias{readDataVersions}
\title{Read Data Versions}
\usage{
readDataVersions(file)
}
\arguments{
\item{file}{Data versions CSV file.}
}
\value{
\code{tbl_df}.
}
\description{
Read Data Versions
}
\note{
The \code{data_versions.csv} file is only generated for special genomes
containing additional information (e.g. the built-in "hg38" build).
}
\examples{
readDataVersions("http://bcbiobase.seq.cloud/data_versions.csv") \%>\%
    glimpse()
}
\seealso{
Other Read Functions: \code{\link{readLogFile}},
  \code{\link{readProgramVersions}},
  \code{\link{readSampleMetadataFile}}
}
\author{
Michael Steinbaugh
}
