% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataVersions.R
\name{readDataVersions}
\alias{readDataVersions}
\title{Read Data Versions}
\usage{
readDataVersions(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
\code{tbl_df}.
}
\description{
Read Data Versions
}
\note{
The \code{data_versions.csv} file is only generated for special genomes
containing additional information (e.g. the built-in "hg38" build).
}
\examples{
file <- file.path(bcbioBaseCacheURL, "data_versions.csv")
x <- readDataVersions(file)
print(x)
}
\seealso{
Other Read Functions: \code{\link{readLog}},
  \code{\link{readProgramVersions}},
  \code{\link{readSampleData}}, \code{\link{readTx2gene}},
  \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
\concept{Read Functions}
