% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{readLog}
\alias{readLog}
\title{Read Log File}
\usage{
readLog(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
\code{character}.
}
\description{
Read Log File
}
\examples{
file <- file.path(bcbioBaseCacheURL, "bcbio_nextgen.log")
x <- readLog(file)
head(x)
}
\seealso{
Other Import/Export Functions: \code{\link{copyToDropbox}},
  \code{\link{readDataVersions}},
  \code{\link{readProgramVersions}},
  \code{\link{readSampleData}}, \code{\link{readTx2Gene}},
  \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
\concept{Import/Export Functions}
