% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{readLog}
\alias{readLog}
\title{Read Log File}
\usage{
readLog(file)
}
\arguments{
\item{file}{File path.}
}
\value{
\code{character}.
}
\description{
Read Log File
}
\examples{
readLog("http://bcbiobase.seq.cloud/bcbio-nextgen.log") \%>\%
    head()
}
\seealso{
Other Read Functions: \code{\link{readDataVersions}},
  \code{\link{readProgramVersions}},
  \code{\link{readSampleData}}, \code{\link{readTx2gene}},
  \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
