% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLogFile.R
\name{readLogFile}
\alias{readLogFile}
\title{Read Log File}
\usage{
readLogFile(file, quiet = FALSE)
}
\arguments{
\item{file}{Log file.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Character vector.
}
\description{
Read Log File
}
\examples{
url <- file.path(
    "http://bcbiobase.seq.cloud",
    "bcbio",
    "bcbio-nextgen.log")
readLogFile(url) \%>\% head()
}
\seealso{
Other File Utilities: \code{\link{readDataVersions}}
}
