% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLogFile.R
\name{readLogFile}
\alias{readLogFile}
\title{Read Log File}
\usage{
readLogFile(file)
}
\arguments{
\item{file}{Log file.}
}
\value{
Character vector.
}
\description{
Read Log File
}
\examples{
readLogFile("http://bcbiobase.seq.cloud/bcbio-nextgen.log") \%>\%
    head()
}
\seealso{
Other Read Functions: \code{\link{readDataVersions}},
  \code{\link{readProgramVersions}},
  \code{\link{readSampleMetadataFile}}
}
\author{
Michael Steinbaugh
}
