% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLogFile.R
\name{readLogFile}
\alias{readLogFile}
\title{Read Log File}
\usage{
readLogFile(file)
}
\arguments{
\item{file}{Log file.}
}
\value{
Character vector.
}
\description{
Read Log File
}
\examples{
url <- paste(
    "http://bcbiobase.seq.cloud",
    "bcbio",
    "bcbio-nextgen.log",
    sep = "/"
)
readLogFile(url) \%>\% head()
}
\seealso{
Other Read Functions: \code{\link{readDataVersions}},
  \code{\link{readProgramVersions}},
  \code{\link{readSampleMetadataFile}}
}
