% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProgramVersions.R
\name{readProgramVersions}
\alias{readProgramVersions}
\title{Read Program Versions}
\usage{
readProgramVersions(file)
}
\arguments{
\item{file}{Program versions TXT file.}
}
\value{
\link{data.frame}.
}
\description{
Read Program Versions
}
\note{
bcbio doesn't save program versions when run in fast mode.
}
\examples{
url <- paste(
    "http://bcbiobase.seq.cloud",
    "bcbio",
    "programs.txt",
    sep = "/"
)
readProgramVersions(url)
}
\seealso{
Other Read Functions: \code{\link{readDataVersions}},
  \code{\link{readLogFile}},
  \code{\link{readSampleMetadataFile}}
}
