% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readProgramVersions.R
\docType{methods}
\name{readProgramVersions}
\alias{readProgramVersions}
\alias{readProgramVersions}
\alias{readProgramVersions,character-method}
\title{Read Program Versions}
\usage{
readProgramVersions(object, ...)

\S4method{readProgramVersions}{character}(object, quiet = FALSE)
}
\arguments{
\item{object}{Project directory path (character vector).}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
\link{data.frame}.
}
\description{
Read Program Versions
}
\examples{
url <- file.path(
    "http://bcbiobase.seq.cloud",
    "bcbio",
    "programs.txt")
readProgramVersions(url)
}
\seealso{
Other bcbio Utilities: \code{\link{prepareSummarizedExperiment}},
  \code{\link{readDataVersions}}
}
\keyword{internal}
