% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProgramVersions.R
\name{readProgramVersions}
\alias{readProgramVersions}
\title{Read Program Versions}
\usage{
readProgramVersions(file, quiet = FALSE)
}
\arguments{
\item{file}{Program versions TXT file.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
\link{data.frame}.
}
\description{
Read Program Versions
}
\examples{
url <- file.path(
    "http://bcbiobase.seq.cloud",
    "bcbio",
    "programs.txt")
readProgramVersions(url)
}
\seealso{
Other bcbio Utilities: \code{\link{prepareSummarizedExperiment}}
}
\keyword{internal}
