% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProgramVersions.R
\name{readProgramVersions}
\alias{readProgramVersions}
\title{Read program versions}
\usage{
readProgramVersions(file)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{DataFrame}.
}
\description{
Read program versions
}
\note{
bcbio doesn't save program versions when run in fast mode.
}
\examples{
file <- file.path(bcbioBaseCacheURL, "programs.txt")
x <- readProgramVersions(file)
print(x)
}
\author{
Michael Steinbaugh
}
