% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readProgramVersions}
\alias{readProgramVersions}
\title{Read Program Versions}
\usage{
readProgramVersions(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
\code{DataFrame}.
}
\description{
Read Program Versions
}
\note{
bcbio doesn't save program versions when run in fast mode.
}
\examples{
file <- file.path(bcbioBaseCacheURL, "programs.txt")
x <- readProgramVersions(file)
print(x)
}
\author{
Michael Steinbaugh
}
