% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProgramVersions.R
\name{readProgramVersions}
\alias{readProgramVersions}
\title{Read Program Versions}
\usage{
readProgramVersions(file)
}
\arguments{
\item{file}{File path.}
}
\value{
\code{tbl_df}.
}
\description{
Read Program Versions
}
\note{
bcbio doesn't save program versions when run in fast mode.
}
\examples{
readProgramVersions("http://bcbiobase.seq.cloud/programs.txt")
}
\seealso{
Other Read Functions: \code{\link{readDataVersions}},
  \code{\link{readLog}}, \code{\link{readSampleData}},
  \code{\link{readTx2gene}}, \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
