% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSampleMetadataFile.R
\name{readSampleMetadataFile}
\alias{readSampleMetadataFile}
\title{Read Sample Metadata File}
\usage{
readSampleMetadataFile(file, lanes = 1L, quiet = FALSE)
}
\arguments{
\item{file}{Metadata file. Supports CSV and XLSX file formats.}

\item{lanes}{\emph{Optional}. Number of lanes used to split the samples into
technical replicates (\code{_LXXX}) suffix.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
\link{data.frame}.
}
\description{
Read Sample Metadata File
}
\examples{
# Demultiplexed
demultiplexed <- file.path(
    "http://bcbiobase.seq.cloud",
    "sample_metadata",
    "demultiplexed.xlsx")
readSampleMetadataFile(demultiplexed) \%>\% glimpse()

# Multiplexed (e.g. inDrop single-cell RNA-seq)
multiplexed <- file.path(
    "http://bcbiobase.seq.cloud",
    "sample_metadata",
    "multiplexed.xlsx")
readSampleMetadataFile(multiplexed) \%>\% glimpse()
}
