% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSampleMetadataFile.R
\name{readSampleMetadataFile}
\alias{readSampleMetadataFile}
\title{Read Sample Metadata File}
\usage{
readSampleMetadataFile(file, lanes = 1L)
}
\arguments{
\item{file}{File path. Supports CSV, TSV, and XLSX file formats.}

\item{lanes}{Number of lanes used to split the samples into technical
replicates (\code{_LXXX}) suffix.}
}
\value{
\code{data.frame}.
}
\description{
Read Sample Metadata File
}
\examples{
# Demultiplexed
readSampleMetadataFile("http://bcbiobase.seq.cloud/demultiplexed.csv") \%>\%
    glimpse()

# Multiplexed (e.g. inDrop single-cell RNA-seq)
readSampleMetadataFile("http://bcbiobase.seq.cloud/multiplexed.csv") \%>\%
    glimpse()
}
\seealso{
Other Read Functions: \code{\link{readDataVersions}},
  \code{\link{readLogFile}},
  \code{\link{readProgramVersions}},
  \code{\link{readTx2gene}}
}
\author{
Michael Steinbaugh
}
