% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSampleMetadataFile.R
\name{readSampleMetadataFile}
\alias{readSampleMetadataFile}
\title{Read Sample Metadata File}
\usage{
readSampleMetadataFile(file, lanes = 1L)
}
\arguments{
\item{file}{Metadata file. Supports CSV and XLSX file formats.}

\item{lanes}{\emph{Optional}. Number of lanes used to split the samples into
technical replicates (\code{_LXXX}) suffix.}
}
\value{
\code{data.frame}.
}
\description{
Read Sample Metadata File
}
\examples{
# Demultiplexed
demultiplexed <- paste(
    "http://bcbiobase.seq.cloud",
    "sample_metadata",
    "demultiplexed.xlsx",
    sep = "/"
)
readSampleMetadataFile(demultiplexed) \%>\% glimpse()

# Multiplexed (e.g. inDrop single-cell RNA-seq)
multiplexed <- paste(
    "http://bcbiobase.seq.cloud",
    "sample_metadata",
    "multiplexed.xlsx",
    sep = "/"
)
readSampleMetadataFile(multiplexed) \%>\% glimpse()
}
\seealso{
Other Read Functions: \code{\link{readDataVersions}},
  \code{\link{readLogFile}},
  \code{\link{readProgramVersions}}
}
