% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readTx2Gene}
\alias{readTx2Gene}
\title{Read Transcript-to-Gene Annotations}
\usage{
readTx2Gene(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
\code{Tx2Gene}.
}
\description{
Generates a \code{Tx2Gene} object containing \code{transcriptID} and \code{geneID} columns.
}
\note{
Doesn't attempt to strip transcript versions.
}
\examples{
file <- file.path(bcbioBaseCacheURL, "tx2gene.csv")
x <- readTx2Gene(file)
print(x)
}
\author{
Michael Steinbaugh
}
