% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTx2Gene.R
\name{readTx2Gene}
\alias{readTx2Gene}
\title{Transcript-to-Gene Annotations}
\usage{
readTx2Gene(file)
}
\arguments{
\item{file}{\code{string}. File path.}
}
\value{
\code{Tx2Gene}.
}
\description{
Generates a \code{Tx2Gene} object containing \code{transcriptID} and \code{geneID} columns.
Rownames are set to the transcript IDs.
}
\note{
Doesn't attempt to strip transcript versions.
}
\examples{
file <- file.path(bcbioBaseCacheURL, "tx2gene.csv")
x <- readTx2Gene(file)
print(x)
}
\seealso{
Other Import/Export Functions: \code{\link{copyToDropbox}},
  \code{\link{readDataVersions}}, \code{\link{readLog}},
  \code{\link{readProgramVersions}},
  \code{\link{readSampleData}}, \code{\link{readYAML}}
}
\author{
Michael Steinbaugh
}
\concept{Import/Export Functions}
