% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTx2Gene.R
\name{readTx2Gene}
\alias{readTx2Gene}
\title{Read Transcript-to-Gene Annotations}
\usage{
readTx2Gene(file, organism = NULL, genomeBuild = NULL,
  ensemblRelease = NULL)
}
\arguments{
\item{file}{\code{character(1)}. File path.}

\item{organism}{\code{character(1)} or \code{NULL}. Full Latin organism name
(e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)} or \code{NULL}. Genome build assembly name
(e.g. \code{"GRCh38"}).}

\item{ensemblRelease}{\code{integer(1)} or \code{NULL}. Ensembl release version
(e.g. \code{90}).}
}
\value{
\code{Tx2Gene}.
}
\description{
Generates a \code{Tx2Gene} object containing \code{transcriptID} and \code{geneID} columns.
}
\note{
Doesn't attempt to strip transcript versions.
}
\examples{
file <- file.path(bcbioBaseCacheURL, "tx2gene.csv")
x <- readTx2Gene(
    file = file,
    organism = "Mus musculus",
    genomeBuild = "GRCm38",
    ensemblRelease = 90L
)
print(x)
}
\author{
Michael Steinbaugh
}
