% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTx2gene.R
\name{readTx2gene}
\alias{readTx2gene}
\title{Transcript to Gene Annotations}
\usage{
readTx2gene(file)
}
\arguments{
\item{file}{File path.}
}
\value{
\code{data.frame}.
}
\description{
Transcript to Gene Annotations
}
\note{
Doesn't attempt to strip transcript versions.
}
\seealso{
Other Read Functions: \code{\link{readDataVersions}},
  \code{\link{readLogFile}},
  \code{\link{readProgramVersions}},
  \code{\link{readSampleMetadataFile}}
}
\author{
Michael Steinbaugh
}
