% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{importSampleData}
\alias{importTx2Gene}
\alias{metadataBlacklist}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump:reexports]{importSampleData}}, \code{\link[basejump:reexports]{importTx2Gene}}, \code{\link[basejump:reexports]{metadataBlacklist}}}
}}

