% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{separatorBar}
\alias{updateMessage}
\alias{prepareTemplate}
\alias{lanePattern}
\alias{minimalSampleData}
\alias{metadataBlacklist}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{separatorBar}}, \code{\link[basejump]{updateMessage}}, \code{\link[basejump]{prepareTemplate}}, \code{\link[basejump]{lanePattern}}, \code{\link[basejump]{minimalSampleData}}, \code{\link[basejump]{metadataBlacklist}}}
}}

