% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assertFormalInterestingGroups}
\alias{reexports}
\alias{convertGenesToSymbols}
\alias{reexports}
\alias{gene2symbol}
\alias{reexports}
\alias{interestingGroups}
\alias{reexports}
\alias{interestingGroups<-}
\alias{reexports}
\alias{sampleData}
\alias{reexports}
\alias{sampleData<-}
\alias{reexports}
\alias{sampleNames}
\alias{reexports}
\alias{sanitizeSampleData}
\alias{reexports}
\alias{selectSamples}
\alias{reexports}
\alias{uniteInterestingGroups}
\alias{reexports}
\alias{pull}
\alias{reexports}
\alias{GRanges}
\alias{reexports}
\alias{colorRampPalette}
\alias{reexports}
\alias{ggplot}
\alias{reexports}
\alias{scale_color_gradient}
\alias{reexports}
\alias{scale_color_gradientn}
\alias{reexports}
\alias{scale_color_hue}
\alias{reexports}
\alias{scale_color_manual}
\alias{reexports}
\alias{scale_color_viridis_c}
\alias{reexports}
\alias{scale_color_viridis_d}
\alias{reexports}
\alias{scale_colour_gradient}
\alias{reexports}
\alias{scale_colour_gradientn}
\alias{reexports}
\alias{scale_colour_hue}
\alias{reexports}
\alias{scale_colour_manual}
\alias{reexports}
\alias{scale_colour_viridis_c}
\alias{reexports}
\alias{scale_colour_viridis_d}
\alias{reexports}
\alias{scale_fill_hue}
\alias{reexports}
\alias{scale_fill_manual}
\alias{reexports}
\alias{scale_fill_viridis_c}
\alias{reexports}
\alias{scale_fill_viridis_d}
\alias{reexports}
\alias{IRanges}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{brewer.pal}
\alias{reexports}
\alias{DataFrame}
\alias{reexports}
\alias{metadata}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{tibble}
\alias{reexports}
\alias{inferno}
\alias{reexports}
\alias{magma}
\alias{reexports}
\alias{plasma}
\alias{reexports}
\alias{scale_color_viridis}
\alias{reexports}
\alias{scale_colour_viridis}
\alias{reexports}
\alias{scale_fill_viridis}
\alias{reexports}
\alias{viridis}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{assertFormalInterestingGroups}}, \code{\link[basejump]{convertGenesToSymbols}}, \code{\link[basejump]{gene2symbol}}, \code{\link[basejump]{interestingGroups}}, \code{\link[basejump]{interestingGroups<-}}, \code{\link[basejump]{sampleData}}, \code{\link[basejump]{sampleData<-}}, \code{\link[basejump]{sampleNames}}, \code{\link[basejump]{sanitizeSampleData}}, \code{\link[basejump]{selectSamples}}, \code{\link[basejump]{uniteInterestingGroups}}}

  \item{dplyr}{\code{\link[dplyr]{pull}}}

  \item{GenomicRanges}{\code{\link[GenomicRanges]{GRanges}}}

  \item{ggplot2}{\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{scale_color_gradient}}, \code{\link[ggplot2]{scale_color_gradientn}}, \code{\link[ggplot2]{scale_color_hue}}, \code{\link[ggplot2]{scale_color_manual}}, \code{\link[ggplot2]{scale_color_viridis_c}}, \code{\link[ggplot2]{scale_color_viridis_d}}, \code{\link[ggplot2]{scale_colour_gradient}}, \code{\link[ggplot2]{scale_colour_gradientn}}, \code{\link[ggplot2]{scale_colour_hue}}, \code{\link[ggplot2]{scale_colour_manual}}, \code{\link[ggplot2]{scale_colour_viridis_c}}, \code{\link[ggplot2]{scale_colour_viridis_d}}, \code{\link[ggplot2]{scale_fill_hue}}, \code{\link[ggplot2]{scale_fill_manual}}, \code{\link[ggplot2]{scale_fill_viridis_c}}, \code{\link[ggplot2]{scale_fill_viridis_d}}}

  \item{grDevices}{\code{\link[grDevices]{colorRampPalette}}}

  \item{IRanges}{\code{\link[IRanges]{IRanges}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{RColorBrewer}{\code{\link[RColorBrewer]{brewer.pal}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}, \code{\link[S4Vectors]{metadata}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{tibble}}}

  \item{viridis}{\code{\link[viridis]{inferno}}, \code{\link[viridis]{magma}}, \code{\link[viridis]{plasma}}, \code{\link[viridis]{scale_color_viridis}}, \code{\link[viridis]{scale_colour_viridis}}, \code{\link[viridis]{scale_fill_viridis}}, \code{\link[viridis]{viridis}}}
}}

