% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pull}
\alias{reexports}
\alias{GRanges}
\alias{reexports}
\alias{ggplot}
\alias{reexports}
\alias{scale_color_hue}
\alias{reexports}
\alias{scale_color_manual}
\alias{reexports}
\alias{scale_colour_hue}
\alias{reexports}
\alias{scale_colour_manual}
\alias{reexports}
\alias{scale_fill_hue}
\alias{reexports}
\alias{scale_fill_manual}
\alias{reexports}
\alias{IRanges}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{DataFrame}
\alias{reexports}
\alias{metadata}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{tibble}
\alias{reexports}
\alias{inferno}
\alias{reexports}
\alias{magma}
\alias{reexports}
\alias{plasma}
\alias{reexports}
\alias{viridis}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{pull}}}

  \item{GenomicRanges}{\code{\link[GenomicRanges]{GRanges}}}

  \item{ggplot2}{\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{scale_color_hue}}, \code{\link[ggplot2]{scale_color_manual}}, \code{\link[ggplot2]{scale_colour_hue}}, \code{\link[ggplot2]{scale_colour_manual}}, \code{\link[ggplot2]{scale_fill_hue}}, \code{\link[ggplot2]{scale_fill_manual}}}

  \item{IRanges}{\code{\link[IRanges]{IRanges}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}, \code{\link[S4Vectors]{metadata}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{tibble}}}

  \item{viridis}{\code{\link[viridis]{inferno}}, \code{\link[viridis]{magma}}, \code{\link[viridis]{plasma}}, \code{\link[viridis]{viridis}}}
}}

