% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{readYAML}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{DataFrame}
\alias{reexports}
\alias{metadata}
\alias{reexports}
\alias{glimpse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{readYAML}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}, \code{\link[S4Vectors]{metadata}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}
}}

