% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{prepareTemplate}
\alias{lanePattern}
\alias{minimalSampleData}
\alias{import}
\alias{pull}
\alias{\%>\%}
\alias{glimpse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{prepareTemplate}}, \code{\link[basejump]{lanePattern}}, \code{\link[basejump]{minimalSampleData}}, \code{\link[basejump]{import}}}

  \item{dplyr}{\code{\link[dplyr]{pull}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}
}}

