% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{minimalSampleData}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{interestingGroups}}, \code{\link[basejump]{interestingGroups<-}}, \code{\link[basejump]{minimalSampleData}}}
}}

