% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{metadataBlacklist}
\alias{readSampleData}
\alias{readTx2Gene}
\alias{importSampleData}
\alias{importTx2Gene}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{importSampleData}}, \code{\link[basejump:reexports]{importTx2Gene}}, \code{\link[basejump:reexports]{metadataBlacklist}}, \code{\link[basejump:deprecated]{readSampleData}}, \code{\link[basejump:deprecated]{readTx2Gene}}}
}}

