% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assertFormalInterestingGroups}
\alias{convertGenesToSymbols}
\alias{gene2symbol}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleNames}
\alias{sanitizeSampleData}
\alias{selectSamples}
\alias{uniteInterestingGroups}
\alias{separatorBar}
\alias{updateMessage}
\alias{prepareTemplate}
\alias{lanePattern}
\alias{minimalSampleData}
\alias{import}
\alias{pull}
\alias{\%>\%}
\alias{glimpse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{assertFormalInterestingGroups}}, \code{\link[basejump]{convertGenesToSymbols}}, \code{\link[basejump]{gene2symbol}}, \code{\link[basejump]{interestingGroups}}, \code{\link[basejump]{interestingGroups<-}}, \code{\link[basejump]{sampleData}}, \code{\link[basejump]{sampleData<-}}, \code{\link[basejump]{sampleNames}}, \code{\link[basejump]{sanitizeSampleData}}, \code{\link[basejump]{selectSamples}}, \code{\link[basejump]{uniteInterestingGroups}}, \code{\link[basejump]{separatorBar}}, \code{\link[basejump]{updateMessage}}, \code{\link[basejump]{prepareTemplate}}, \code{\link[basejump]{lanePattern}}, \code{\link[basejump]{minimalSampleData}}, \code{\link[basejump]{import}}}

  \item{dplyr}{\code{\link[dplyr]{pull}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}
}}

