% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{prepareTemplate}
\alias{pull}
\alias{ggplot}
\alias{\%>\%}
\alias{glimpse}
\alias{tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{prepareTemplate}}}

  \item{dplyr}{\code{\link[dplyr]{pull}}}

  \item{ggplot2}{\code{\link[ggplot2]{ggplot}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{tibble}}}
}}

