% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{annotable}
\alias{reexports}
\alias{camel}
\alias{reexports}
\alias{checkGene2symbol}
\alias{reexports}
\alias{detectOrganism}
\alias{reexports}
\alias{fixNA}
\alias{reexports}
\alias{gene2symbol}
\alias{reexports}
\alias{plotHeatmap}
\alias{reexports}
\alias{readYAML}
\alias{reexports}
\alias{sanitizeAnnotable}
\alias{reexports}
\alias{snake}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{DataFrame}
\alias{reexports}
\alias{metadata}
\alias{reexports}
\alias{glimpse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{annotable}}, \code{\link[basejump]{camel}}, \code{\link[basejump]{checkGene2symbol}}, \code{\link[basejump]{detectOrganism}}, \code{\link[basejump]{fixNA}}, \code{\link[basejump]{gene2symbol}}, \code{\link[basejump]{plotHeatmap}}, \code{\link[basejump]{readYAML}}, \code{\link[basejump]{sanitizeAnnotable}}, \code{\link[basejump]{snake}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}, \code{\link[S4Vectors]{metadata}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}
}}

