% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{metadataBlacklist}
\alias{readSampleData}
\alias{readTx2Gene}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{metadataBlacklist}}, \code{\link[basejump]{readSampleData}}, \code{\link[basejump]{readTx2Gene}}}
}}

