% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{annotable}
\alias{reexports}
\alias{aggregateReplicates}
\alias{reexports}
\alias{assignAndSaveData}
\alias{reexports}
\alias{camel}
\alias{reexports}
\alias{checkAnnotable}
\alias{reexports}
\alias{checkGene2symbol}
\alias{reexports}
\alias{convertGenesToSymbols}
\alias{reexports}
\alias{convertTranscriptsToGenes}
\alias{reexports}
\alias{detectOrganism}
\alias{reexports}
\alias{dynamicPlotlist}
\alias{reexports}
\alias{fixNA}
\alias{reexports}
\alias{gene2symbol}
\alias{reexports}
\alias{gene2symbolFromGTF}
\alias{reexports}
\alias{loadData}
\alias{reexports}
\alias{loadDataAsName}
\alias{reexports}
\alias{loadRemoteData}
\alias{reexports}
\alias{mdHeader}
\alias{reexports}
\alias{mdList}
\alias{reexports}
\alias{midnightTheme}
\alias{reexports}
\alias{plotHeatmap}
\alias{reexports}
\alias{plotQuantileHeatmap}
\alias{reexports}
\alias{readFileByExtension}
\alias{reexports}
\alias{readGTF}
\alias{reexports}
\alias{readYAML}
\alias{reexports}
\alias{sanitizeAnnotable}
\alias{reexports}
\alias{saveData}
\alias{reexports}
\alias{snake}
\alias{reexports}
\alias{tx2gene}
\alias{reexports}
\alias{tx2geneFromGTF}
\alias{reexports}
\alias{writeCounts}
\alias{reexports}
\alias{pull}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{!!}
\alias{reexports}
\alias{!!!}
\alias{reexports}
\alias{.data}
\alias{reexports}
\alias{abort}
\alias{reexports}
\alias{inform}
\alias{reexports}
\alias{is_string}
\alias{reexports}
\alias{sym}
\alias{reexports}
\alias{syms}
\alias{reexports}
\alias{warn}
\alias{reexports}
\alias{DataFrame}
\alias{reexports}
\alias{metadata}
\alias{reexports}
\alias{glimpse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{annotable}}, \code{\link[basejump]{aggregateReplicates}}, \code{\link[basejump]{assignAndSaveData}}, \code{\link[basejump]{camel}}, \code{\link[basejump]{checkAnnotable}}, \code{\link[basejump]{checkGene2symbol}}, \code{\link[basejump]{convertGenesToSymbols}}, \code{\link[basejump]{convertTranscriptsToGenes}}, \code{\link[basejump]{detectOrganism}}, \code{\link[basejump]{dynamicPlotlist}}, \code{\link[basejump]{fixNA}}, \code{\link[basejump]{gene2symbol}}, \code{\link[basejump]{gene2symbolFromGTF}}, \code{\link[basejump]{loadData}}, \code{\link[basejump]{loadDataAsName}}, \code{\link[basejump]{loadRemoteData}}, \code{\link[basejump]{mdHeader}}, \code{\link[basejump]{mdList}}, \code{\link[basejump]{midnightTheme}}, \code{\link[basejump]{plotHeatmap}}, \code{\link[basejump]{plotQuantileHeatmap}}, \code{\link[basejump]{readFileByExtension}}, \code{\link[basejump]{readGTF}}, \code{\link[basejump]{readYAML}}, \code{\link[basejump]{sanitizeAnnotable}}, \code{\link[basejump]{saveData}}, \code{\link[basejump]{snake}}, \code{\link[basejump]{tx2gene}}, \code{\link[basejump]{tx2geneFromGTF}}, \code{\link[basejump]{writeCounts}}}

  \item{dplyr}{\code{\link[dplyr]{pull}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{!!}}, \code{\link[rlang]{!!!}}, \code{\link[rlang]{.data}}, \code{\link[rlang]{abort}}, \code{\link[rlang]{inform}}, \code{\link[rlang]{is_string}}, \code{\link[rlang]{sym}}, \code{\link[rlang]{syms}}, \code{\link[rlang]{warn}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}, \code{\link[S4Vectors]{metadata}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}
}}

