% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleData.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData,ANY-method}
\alias{sampleData<-,ANY-method}
\title{Sample Data}
\usage{
\S4method{sampleData}{ANY}(object, return = c("DataFrame", "data.frame"), ...)

\S4method{sampleData}{ANY}(object, ...) <- value
}
\arguments{
\item{return}{Return as \code{DataFrame} or \code{data.frame}.}
}
\value{
\code{DataFrame} or \code{data.frame} containing sample metadata. Note that
the samples are defined in the rows, similar to \code{\link[=colData]{colData()}}.
}
\description{
Returns the sample metadata.
}
\note{
This is a complement to the standard \code{\link[=colData]{colData()}} function, but improves
support for accessing sample metadata for datasets where multiple items in
the columns map to a single sample (e.g. cells for a single-cell RNA-seq
experiment).
}
