% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleData.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleData}
\alias{sampleData,SummarizedExperiment-method}
\alias{sampleData<-,SummarizedExperiment-method}
\alias{sampleMetadata}
\alias{sampleMetadata<-}
\title{Sample Data}
\usage{
sampleData(object, ...)

sampleData(object, ...) <- value

\S4method{sampleData}{SummarizedExperiment}(object, return = c("data.frame",
  "DataFrame", "kable"))

\S4method{sampleData}{SummarizedExperiment}(object, ...) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}
}
\value{
Data describing the samples.
}
\description{
Return the sample metadata.
}
\note{
This is a complement to the standard \code{\link[=colData]{colData()}} function, but improves
support for accessing sample metadata for datasets where multiple items in
the columns map to a single sample (e.g. cells for a single-cell RNA-seq
experiment).
}
\author{
Michael Steinbaugh
}
