% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDirs.R
\name{sampleDirs}
\alias{sampleDirs}
\title{Sample Directories}
\usage{
sampleDirs(uploadDir)
}
\arguments{
\item{uploadDir}{bcbio run upload directory.}
}
\value{
Named character vector containing sample directory paths. Function
will abort if no sample directories match.
}
\description{
Sample Directories
}
\seealso{
Other Data Functions: \code{\link{gene2symbol}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
