% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleNames.R
\name{sampleNames}
\alias{sampleNames}
\title{Sample Names}
\value{
Named \code{character} vector of the sample names.
}
\description{
This function will return the human readable sample names if defined
in the \code{sampleName} column of \code{\link[=sampleData]{sampleData()}}. Otherwise it will return
the syntactically valid names defined as the rownames of \code{\link[=sampleData]{sampleData()}}.
}
\examples{
# SummarizedExperiment ====
sampleNames(rse_bcb)
sampleNames(rse_dds)
}
\seealso{
Other Metadata Functions: \code{\link{sampleData}}
}
\author{
Michael Steinbaugh
}
