% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleYAML.R
\docType{methods}
\name{sampleYAML}
\alias{sampleYAML}
\alias{sampleYAMLMetadata}
\alias{sampleYAMLMetrics}
\alias{sampleYAML}
\alias{sampleYAML,list,character-method}
\alias{sampleYAMLMetadata,list-method}
\alias{sampleYAMLMetrics,list-method}
\title{Sample YAML Metadata}
\usage{
sampleYAML(yaml, keys, ...)

sampleYAMLMetadata(yaml, ...)

sampleYAMLMetrics(yaml, ...)

\S4method{sampleYAML}{list,character}(yaml, keys)

\S4method{sampleYAMLMetadata}{list}(yaml)

\S4method{sampleYAMLMetrics}{list}(yaml)
}
\arguments{
\item{yaml}{Project summary YAML \code{list}.}

\item{keys}{Nested operator keys, supplied as a character vector.}

\item{...}{Additional arguments.}
}
\value{
\code{tbl_df}.
}
\description{
Sample YAML Metadata
}
\note{
Metrics are only generated for a standard RNA-seq run with aligned
counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't
output the same metrics into the YAML.
}
\examples{
yaml <- basejump::readYAML(
    "http://bcbiobase.seq.cloud/project-summary.yaml"
)

sampleYAML(yaml, keys = "metadata") \%>\% glimpse()
sampleYAMLMetadata(yaml) \%>\% glimpse()
sampleYAMLMetrics(yaml) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
