% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleYAMLMetadata.R
\docType{methods}
\name{sampleYAMLMetadata}
\alias{sampleYAMLMetadata}
\alias{sampleYAMLMetadata}
\alias{sampleYAMLMetadata,list-method}
\title{Sample Metadata from YAML}
\usage{
sampleYAMLMetadata(yaml, ...)

\S4method{sampleYAMLMetadata}{list}(yaml)
}
\arguments{
\item{yaml}{Project summary YAML list.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{data.frame}.
}
\description{
Sample Metadata from YAML
}
\examples{
url <- file.path(
    "http://basejump.seq.cloud",
    "bcbio",
    "project-summary.yaml")
yaml <- basejump::readYAML(url)
sampleYAMLMetadata(yaml)
}
\seealso{
Other YAML Utilities: \code{\link{sampleYAMLMetrics}},
  \code{\link{sampleYAML}}
}
