% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{yaml}
\alias{yaml}
\alias{getGTFFileFromYAML}
\alias{getMetricsFromYAML}
\alias{getSampleDataFromYAML}
\title{YAML Parsing Functions}
\usage{
getGTFFileFromYAML(yaml)

getMetricsFromYAML(yaml)

getSampleDataFromYAML(yaml)
}
\arguments{
\item{yaml}{\code{list}. Project summary YAML.}
}
\value{
\code{string} or \code{DataFrame}.
}
\description{
YAML Parsing Functions
}
\section{Functions}{
\itemize{
\item \code{getGTFFileFromYAML}: \code{string}. GTF file path.

\item \code{getMetricsFromYAML}: \code{DataFrame}. Quality control summary metrics.

\item \code{getSampleDataFromYAML}: \code{DataFrame}. Sample metadata.
}}

\note{
Metrics are only generated for a standard RNA-seq run with aligned
counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't
output the same metrics into the YAML.
}
\examples{
file <- file.path(bcbioBaseCacheURL, "summary.yaml")
yaml <- basejump.io::import(file)
summary(yaml)

## GTF file path.
x <- getGTFFileFromYAML(yaml)
print(x)

## Sample metrics.
x <- getMetricsFromYAML(yaml)
summary(x)
colnames(x)

## Sample metadata.
x <- getSampleDataFromYAML(yaml)
summary(x)
colnames(x)
}
\author{
Michael Steinbaugh
}
