% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{yaml}
\alias{yaml}
\alias{getGTFFileFromYAML}
\alias{getMetricsFromYAML}
\alias{getSampleDataFromYAML}
\title{Get from YAML}
\usage{
getGTFFileFromYAML(yaml)

getMetricsFromYAML(yaml)

getSampleDataFromYAML(yaml)
}
\arguments{
\item{file}{\code{string}. File path to bcbio \code{project-summary.yaml}.}
}
\value{
\code{DataFrame}.
}
\description{
Get from YAML
}
\note{
Metrics are only generated for a standard RNA-seq run with aligned
counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't
output the same metrics into the YAML.
}
\examples{
file <- file.path(bcbioBaseCacheURL, "summary.yaml")

## GTF file path.
x <- getGTFFileFromYAML(yaml)

## Sample metrics.
x <- getMetricsFromYAML(yaml)
print(x)

## Sample metadata.
x <- getSampleDataFromYAML(yaml)
print(x)
}
\author{
Michael Steinbaugh
}
